Defunct {base}	R Documentation

Marking Objects as Defunct

Description

When a function is removed from R it should be replaced by a function which calls .Defunct.

Usage


.Defunct(new, package = NULL, msg)

Arguments

new
character string: A suggestion for a replacement function.

package
character string: The package to be used when suggesting where the defunct function might be listed.

msg
character string: A message to be printed, if missing a default message is used.

Details

.Defunct is called from defunct functions. Functions should be listed in help("pkg-defunct") for an appropriate pkg, including base (with the alias added to the respective Rd file).

.Defunct signals an error of class defunctError with fields old, new, and package.

See Also

Deprecated.

base-defunct and so on which list the defunct functions in the packages.

[Package base version 3.6.1 Index]