Deprecated {base}	R Documentation

Marking Objects as Deprecated

Description

When an object is about to be removed from R it is first deprecated and should include a call to .Deprecated.

Usage

.Deprecated(new, package=NULL, msg,
            old = as.character(sys.call(sys.parent()))[1L])

Arguments

new
character string: A suggestion for a replacement function.

package
character string: The package to be used when suggesting where the deprecated function might be listed.

msg
character string: A message to be printed, if missing a default message is used.

old
character string specifying the function (default) or usage which is being deprecated.

Details

.Deprecated("<new name>") is called from deprecated functions. The original help page for these functions is often available at help("oldName-deprecated") (note the quotes). Functions should be listed in help("pkg-deprecated") for an appropriate pkg, including base.

See Also

Defunct

base-deprecated and so on which list the deprecated functions in the packages.

[Package base version 2.16.0 Index]