Internal {base}	R Documentation

Call an Internal Function

Description

.Internal performs a call to an internal code which is built in to the R interpreter.

Only true R wizards should even consider using this function, and only R developers can add to the list of internal functions.

Usage


 .Internal(call)

Arguments

call
a call expression

See Also

.Primitive, .External (the nearest equivalent available to users).

[Package base version 3.6.1 Index]