.Platform {base}	R Documentation

Platform Specific Variables

Description

.Platform is a list with some details of the platform under which R was built. This provides means to write OS portable R code.

Usage

.Platform

Value

A list with at least the following components:

OS.type 	 character, giving the Operating System (family) of the computer. One of "unix" or "windows".

file.sep 	 character, giving the file separator, used on your platform, e.g., "/" on Unix alikes.

dynlib.ext 	 character, giving the file name extension of dynamically loadable libraries, e.g., ".dll" on Windows.

GUI 	 character, giving the type of GUI in use, or "unknown" if no GUI can be assumed.

endian 	 character, "big" or "little", giving the endianness of the processor in use.

pkgType 	 character, the preferred setting for options("pkgType"). Values "source", "mac.binary" and "win.binary" are currently in use.

See Also

R.version and Sys.info give more details about the OS. In particular, R.version$platform is the canonical name of the platform under which R was compiled.

.Machine for details of the arithmetic used, and system for invoking platform-specific system commands.

Examples

## Note: this can be done in a system-independent way by file.info()$isdir
if(.Platform$OS.type == "unix") {
   system.test <- function(...) { system(paste("test", ...)) == 0 }
   dir.exists <- function(dir) sapply(dir, function(d)system.test("-d", d))
   dir.exists(c(R.home(), "/tmp", "~", "/NO"))# > T T T F
}

[Package base version 2.1.0 Index]