Startup {base}	R Documentation

Initialization at Start of an R Session

Description

In R, the startup mechanism is as follows.

Unless --no-environ was given on the command line, R searches for user and site files to process for setting environment variables. The name of the site file is the one pointed to by the environment variable R_ENVIRON; if this is unset or empty, ‘$R_HOME/etc/Renviron.site’ is used (if it exists, which it does not in a “factory-fresh” installation). The user files searched for are ‘.Renviron’ in the current or in the user's home directory (in that order). See Details for how the files are read.

Then R searches for the site-wide startup profile unless the command line option --no-site-file was given. The name of this file is taken from the value of the R_PROFILE environment variable. If this variable is unset, the default is ‘$R_HOME/etc/Rprofile.site’, which is used if it exists (which it does not in a “factory-fresh” installation). This code is loaded into package base. Users need to be careful not to unintentionally overwrite objects in base, and it is normally advisable to use local if code needs to be executed: see the examples.

Then, unless --no-init-file was given, R searches for a file called ‘.Rprofile’ in the current directory or in the user's home directory (in that order) and sources it into the user workspace.

It then loads a saved image of the user workspace from ‘.RData’ if there is one (unless --no-restore-data was specified, or --no-restore, on the command line).

Next, if a function .First is found on the search path, it is executed as .First(). Finally, function .First.sys() in the base package is run. This calls require to attach the default packages specified by options("defaultPackages").

A function .First (and .Last) can be defined in appropriate ‘.Rprofile’ or ‘Rprofile.site’ files or have been saved in ‘.RData’. If you want a different set of packages than the default ones when you start, insert a call to options in the ‘.Rprofile’ or ‘Rprofile.site’ file. For example, options(defaultPackages = character()) will attach no extra packages on startup. Alternatively, set R_DEFAULT_PACKAGES=NULL as an environment variable before running R. Using options(defaultPackages = "") or R_DEFAULT_PACKAGES="" enforces the R system default.

The commands history is read from the file specified by the environment variable R_HISTFILE (default ‘.Rhistory’) unless --no-restore-history was specified (or --no-restore).

The command-line flag --vanilla implies --no-site-file, --no-init-file, --no-restore and --no-environ.

Usage

.First <- function() { ...... }

.Rprofile <startup file>

Details

Note that there are two sorts of files used in startup: environment files which contain lists of environment variables to be set, and profile files which contain R code.

Lines in a site or user environment file should be either comment lines starting with #, or lines of the form name=value. The latter sets the environmental variable name to value, overriding an existing value. If value is of the form ${foo-bar}, the value is that of the environmental variable foo if that exists and is set to a non-empty value, otherwise bar. This construction can be nested, so bar can be of the same form (as in ${foo-${bar-blah}}).

Leading and trailing white space in value are stripped. value is processed in a similar way to a Unix shell. In particular quotes are stripped, and backslashes are removed except inside quotes.

Historical notes

Prior to R version 1.4.0, the environment files searched were ‘.Renviron’ in the current directory, the file pointed to by R_ENVIRON if set, and ‘.Renviron’ in the user's home directory.

Prior to R version 1.2.1, ‘.Rprofile’ was sourced after ‘.RData’ was loaded, although the documented order was as here.

The format for site and user environment files was changed in version 1.2.0. Older files are quite likely to work but may generate warnings on startup if they contained unnecessary export statements.

Values in environment files were not processed prior to version 1.4.0.

See Also

.Last for final actions before termination.

For profiling code, see Rprof.

Examples

## Not run: 
# Example ~/.Renviron on Unix
R_LIBS=~/R/library
PAGER=/usr/local/bin/less

# Example .Renviron on Windows
R_LIBS=C:/R/library
MY_TCLTK=yes
TCL_LIBRARY=c:/packages/Tcl/lib/tcl8.4

# Example of .Rprofile
options(width=65, digits=5)
options(show.signif.stars=FALSE)
ps.options(horizontal=FALSE)
set.seed(1234)
.First <- function() cat("\n   Welcome to R!\n\n")
.Last <- function()  cat("\n   Goodbye!\n\n")

# Example of Rprofile.site
local({
  old <- getOption("defaultPackages")
  options(defaultPackages = c(old, "MASS"))
})

## if .Renviron contains
FOOBAR="coo\bar"doh\ex"abc\"def'"

## then we get
> cat(Sys.getenv("FOOBAR"), "\n")
coo\bardoh\exabc"def'
## End(Not run)

[Package base version 2.1.0 Index]