Sys.time {base}	R Documentation

Get Current Date, Time and Timezone

Description

Sys.time and Sys.Date returns the system's idea of the current date with and without time, and Sys.timezone returns the current time zone.

Usage

Sys.time()
Sys.Date()
Sys.timezone()

Details

Sys.time returns an absolute date-time value which can be converted in various time zones and may return different days.

Sys.Date returns the day in the current timezone.

Value

Sys.time returns an object of class "POSIXct" (see DateTimeClasses).
Sys.Date returns an object of class "Date" (see Date).
Sys.timezone returns an OS-specific character string, possibly an empty string.

See Also

date for the system time in a fixed-format character string.

Examples

Sys.time()
## locale-specific version of date()
format(Sys.time(), "%a %b %d %X %Y")

Sys.Date()

Sys.timezone()

[Package base version 2.1.0 Index]