Trig {base}	R Documentation

Trigonometric Functions

Description

These functions give the obvious trigonometric functions. They respectively compute the cosine, sine, tangent, arc-cosine, arc-sine, arc-tangent, and the two-argument arc-tangent.

Usage

cos(x)
sin(x)
tan(x)
acos(x)
asin(x)
atan(x)
atan2(y, x)

Arguments

x, y 	 numeric vector

Details

The arc-tangent of two arguments atan2(y,x) returns the angle between the x-axis and the vector from the origin to (x,y), i.e., for positive arguments atan2(y,x) == atan(y/x).

Angles are in radians, not degrees (i.e., a right angle is π/2).

All except atan2 are generic functions: methods can be defined for them individually or via the Math group generic.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

[Package base version 2.1.0 Index]