R.Version {base}	R Documentation

Version Information

Description

R.Version() provides detailed information about the version of R running.
R.version is a variable (a list) holding this information (and version is a copy of it for S compatibility), whereas R.version.string is a simple character string, useful for plotting, etc.

Usage

R.Version()
R.version
R.version.string

Value

R.Version returns a list with components

platform 	 the platform for which R was built. Under Unix, a triplet of the form CPU-VENDOR-OS, as determined by the configure script. E.g, "i586-unknown-linux".

arch 	 the architecture (CPU) R was built on/for.

os 	 the underlying operating system

system 	 CPU and OS.

status 	 the status of the version (e.g., "Alpha")

status.rev 	 the status revision level

major 	 the major version number

minor 	 the minor version number

year 	 the year the version was released

month 	 the month the version was released

day 	 the day the version was released

language 	 always "R".

Note

Do not use R.version$os to test the platform the code is running on: use .Platform$OS.type instead. Slightly different versions of the OS may report different values of R.version$os, as may different versions of R.

See Also

getRversion .Platform.

Examples

R.version$os # to check how lucky you are ...
plot(0) # any plot
mtext(R.version.string, side=1,line=4,adj=1)# a useful bottom-right note

[Package base version 2.1.0 Index]