attr {base}	R Documentation

Object Attributes

Description

Get or set specific attributes of an object.

Usage

attr(x, which)
attr(x, which) <- value

Arguments

x 	 an object whose attributes are to be accessed.

which 	 a character string specifying which attribute is to be accessed.

value 	 an object, the new value of the attribute.

Value

This function provides access to a single object attribute. The simple form above returns the value of the named attribute. The assignment form causes the named attribute to take the value on the right of the assignment symbol.
The first form first looks for an exact match to code amongst the attributed of x, then a partial match. If no exact match is found and more than one partial match is found, the result is NULL.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

attributes

Examples

# create a 2 by 5 matrix
x <- 1:10
attr(x,"dim") <- c(2, 5)

[Package base version 2.1.0 Index]