base-defunct {base}	R Documentation

Defunct Functions in Base Package

Description

The functions or variables listed here are no longer part of R as they are not needed (any more).

Usage

Version()
provide(package)
.Provided
category(...)
dnchisq(.)
pnchisq(.)
qnchisq(.)
rnchisq(.)
print.anova.glm(.)
print.anova.lm(.)
print.tabular(.)
print.plot(.)
save.plot(.)
system.test(.)
dotplot(...)
stripplot(...)
getenv(...)
read.table.url(url, method,...)
scan.url(url, file = tempfile(), method, ...)
source.url(url, file = tempfile(), method, ...)
httpclient(url, port=80, error.is.fatal=TRUE, check.MIME.type=TRUE,
           file=tempfile(), drop.ctrl.z=TRUE)
parse.dcf(text = NULL, file = "", fields = NULL, versionfix = FALSE)
.Alias(expr)
reshapeWide(x, i, j, val, jnames = levels(j))
reshapeLong(x,jvars,  ilev = row.names(x),
            jlev = names(x)[jvars], iname = "reshape.i",
            jname = "reshape.j", vname = "reshape.v")
piechart(x, labels = names(x), edges = 200, radius = 0.8,
         density = NULL, angle = 45, col = NULL, main = NULL, ...)
print.ordered(.)
.Dyn.libs
.lib.loc
machine()
Machine()
Platform()
restart()
printNoClass(x, digits = NULL, quote = TRUE, na.print = NULL,
             print.gap = NULL, right = FALSE, ...)
plot.mts(x, plot.type = c("multiple", "single"), panel = lines,
         log = "", col = par("col"),  bg = NA, pch = par("pch"),
         cex = par("cex"), lty = par("lty"), lwd = par("lwd"),
         ann = par("ann"),  xlab = "Time", type = "l", main=NULL,
         oma=c(6, 0, 5, 0), ...)
print.coefmat(x, digits=max(3, getOption("digits") - 2),
              signif.stars = getOption("show.signif.stars"),
              dig.tst = max(1, min(5, digits - 1)),
              cs.ind, tst.ind, zap.ind = integer(0),
              P.values = NULL, has.Pvalue, eps.Pvalue = .Machine$double.eps,
              na.print = "", ...)

codes(x, ...)
codes(x, ...) <- value

anovalist.lm(object, ..., test = NULL)
lm.fit.null(x, y, method = "qr", tol = 1e-07, ...)
lm.wfit.null(x, y, w, method = "qr", tol = 1e-07, ...)
glm.fit.null(x, y, weights, start = NULL,
             etastart = NULL, mustart = NULL, offset,
             family = gaussian(), control = glm.control(),
             intercept = FALSE)
print.atomic(x, quote = TRUE, ...)
La.eigen(x, symmetric, only.values = FALSE, method = c("dsyevr", "dsyev"))
tetragamma(x)
pentagamma(x)
package.description(pkg, lib.loc = NULL, fields = NULL)

Details

category has been an old-S function before there were factors; should be replaced by factor throughout!

The *chisq() functions now take an optional non-centrality argument, so the *nchisq() functions are no longer needed.

The new function dev.print() should now be used for saving plots to a file or printing them.

provide and its object .Provided have been removed. They were never used for their intended purpose, to allow one package to subsume another.

dotplot and stripplot have been renamed to dotchart and stripchart, respectively.

getenv has been replaced by Sys.getenv.

*.url are replaced by calling read.table, scan or source on a url connection.

httpclient was used by the deprecated "socket" method of download.file.

parse.dcf has been replaced by read.dcf, which is much faster, but has a slightly different interface.

.Alias provided an unreliable way to create duplicate references to the same object. There is no direct replacement. Where multiple references to a single object are required for semantic reasons consider using environments or external pointers. There are some notes on http://developer.r-project.org.

reshape*, which were experimental, are replaced by reshape. This has a different syntax and allows multiple time-varying variables.

piechart is the old name for pie, but clashed with usage in Trellis.

.Dyn.libs and .lib.loc were internal variables used for storing and manipulating the information about packages with dynloaded shared libs, and the known R library trees. These are now dynamic variables which one can get or set using .dynLibs and .libPaths, respectively.

Machine() and Platform() were functions returning the variables .Machine and .Platform respectively.

restart() should be replaced by try(), in preparation for an exception-based implementation. If you use restart() in a way that cannot be replaced with try() then ask for help on r-devel.

printNoClass was in package methods and calls directly the internal function print.default.

plot.mts has been removed, as plot.ts now has the same functionality.

print.coefmat was an older name for printCoefmat with a different default for na.print.

codes was almost always used inappropriately. To get the internal coding of a factor, use unclass, as.vector or as.integer. For ordered factors, codes was equivalent to these, but for unordered factors it assumed an an alphabetical ordering of the levels in the locale in use.

anovalist.lm was replaced by anova.lmlist in R 1.2.0.

lm.fit.null and lm.wfit.null are superseded by lm.fit and lm.wfit which handle null models now. Similarly, glm.fit.null is superseded by glm.fit.

print.atomic differed from print.default only in its argument sequence. It is not a method for print.

La.eigen has become the default for eigen.

tetragamma and pentagamma have been replaced by psigamma.

package.description has been replaced by packageDescription.

See Also

Defunct, base-deprecated

[Package base version 2.1.0 Index]