base-deprecated {base}	R Documentation

Deprecated Functions in Base package

Description

These functions are provided for compatibility with older versions of R only, and may be defunct as soon as the next release.

Usage

loadURL(url, envir = parent.frame(), quiet = TRUE, ...)
delay(x, env = .GlobalEnv)

Arguments

url 	 a character string naming a URL.

envir 	 the environment where the data should be loaded.

quiet, ... 	 additional arguments to download.file.

x 	 an expression

env 	 an evaluation environment

Details

The original help page for these functions is often available at help("oldName-deprecated") (note the quotes). Functions in packages other than the base package are listed in help("pkg-deprecated").

loadURL has been superseded by load(url()). Note the comments on that help page: Windows users will need to use mode="wb".

delay has been replaced by delayedAssign. As from R 2.1.0 promises should never be visible unevaluated.

See Also

Deprecated

[Package base version 2.1.0 Index]