base-internal {base}	R Documentation

Internal Functions in the Base Package

Description

Internal functions in the base package, which are only user-visible because of the special nature of the base namespace.

Usage

.subset(x, ...)
.subset2(x, ...)

.find.package(package, lib.loc = NULL, quiet = FALSE,
              verbose = getOption("verbose"))

.path.package(package, quiet = FALSE)

.getRequiredPackages(file = "DESCRIPTION", quietly = FALSE,
                     useImports = FALSE)
.getRequiredPackages2(pkgInfo, quietly = FALSE, lib.loc = NULL,
                      useImports = FALSE)

testPlatformEquivalence(built, run)

Arguments

x 	 object from which to extract elements.

package 	 the name of a package.

lib.loc 	 a character vector describing the location of R library trees to search through, or NULL. The default value of NULL corresponds to all libraries currently known.

quiet 	 logical. Should this not give warnings or an error if the package(s) are not found?

useImports 	 logical. Should the Imports field be processed?

verbose 	 a logical. If TRUE, additional diagnostics are printed.

file 	 The full path to the ‘DESCRIPTION’ file of a package.

quietly 	 logical: should progress messages be given.

pkginfo 	 A processed ‘DESCRIPTION’ file of class "packageDescription2".

built, run 	 platform strings of the form "cpu-vendor-os".

Details

The functions .subset and .subset2 are essentially equivalent to the [ and [[ operators, except methods that dispatch does not take place. This is to avoid expensive unclassing when applying the default method to an object. They should not normally be invoked by end users.

.find.package returns the paths to the locations where the given packages can be found. If lib.loc is NULL, then then attached packages are searched before the libraries. If a package is found more than once, the first match is used. Unless quiet = TRUE a warning will be given about the named packages which are not found, and an error if none are. If verbose is true, warnings about packages found more than once are given. It does know about versioned installs of packages, and for a package to be returned it must have a ‘DESCRIPTION’ file containing a valid version field.

.path.package returns the paths from which the named packages were loaded, or if none were named, for all currently loaded packages. Unless quiet = TRUE it will warn if some of the packages named are not loaded, and given an error if none are. This function is not meant to be called by users, and its interface might change in future versions. Note that currently package must include the version mumber if versioned installs are used.

.getRequiredPackages2 attaches all the packages mentioned in the Depends field: failure to find a package is an error. It also checks the versions of the packages found against the Depends field.

.getRequiredPackages is a wrapper to .getRequiredPackages2 using a ‘DESCRIPTION’ file.

testPlatformEquivalence is used to test if a package with compiled code can safely be loaded into a build of R, which it does by comparing their ‘canonical’ "cpu-vendor-os" descriptions. The default function omits the vendor part and allows small mismatches elsewhere. Cognescenti can replace this function in the base namespace.

Value

.find.package and .path.packages return a character vector of paths to installed packages.
.getRequiredPackages and .getRequiredPackages2 return invisible().
testPlatformEquivalence: TRUE or FALSE.

[Package base version 2.1.0 Index]