browser {base}	R Documentation

Environment Browser

Description

Interrupt the execution of an expression and allow the inspection of the environment where browser was called from.

Usage

browser()

Details

A call to browser causes a pause in the execution of the current expression and runs a copy of the R interpreter which has access to variables local to the environment where the call took place.

Local variables can be listed with ls, and manipulated with R expressions typed to this sub-interpreter. The sub-interpreter can be exited by typing c. Execution then resumes at the statement following the call to browser.

Typing n causes the step-through-debugger, to start and it is possible to step through the remainder of the function one line at a time. In this mode c will continue to the end of the current context (to the next loop iteration if within a loop).

Typing Q quits the current execution and returns you to the top-level prompt.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Chambers, J. M. (1998) Programming with Data. A Guide to the S Language. Springer.

See Also

debug, and traceback for the stack on error.

[Package base version 2.1.0 Index]