capabilities {base}	R Documentation

Report Capabilities of this Build of R

Description

Report on the optional features which have been compiled into this build of R.

Usage

capabilities(what = NULL)

Arguments

what
character vector or NULL, specifying required components. NULL implies that all are required.

Value

A named logical vector. Current components are

jpeg
is the jpeg function operational?

png
is the png function operational?

tiff
is the tiff function operational?

tcltk
is the tcltk package operational? Note that to make use of Tk you will almost always need to check that "X11" is also available.

X11
Are the X11 graphics device and the X11-based data editor available? This loads the X11 module if not already loaded, and checks that the default display can be contacted unless a X11 device has already been used.

aqua
Are the R.app GUI components and the quartz function operational? Only on some OS X builds. Note that this is distinct from .Platform$GUI == "AQUA", which is true when using the Mac R.app GUI console.

http/ftp
Are url and the internal method for download.file available?

sockets
Are make.socket and related functions available?

libxml
is there support for integrating libxml with the R event loop?

fifo
are FIFO connections supported?

cledit
is command-line editing available in the current R session? This is false in non-interactive sessions. It will be true for the command-line interface if readline support has been compiled in and --no-readline was not used when R was invoked. (If --interactive was used, command-line editing will not actually be available.)

iconv
is internationalization conversion via iconv supported? Always true in current R.

NLS
is there Natural Language Support (for message translations)?

profmem
is there support for memory profiling? See tracemem.

cairo
is there support the svg, cairo_pdf and cairo_ps devices, and for type = "cairo" in the X11, bmp, jpeg, png, and tiff devices?

Note to OS X users

Capabilities "jpeg", "png" and "tiff" refer to the X11-based versions of these devices. If capabilities("aqua") is true, then these devices with type = "quartz" will be available, and out-of-the-box will be the default type. Thus for example the tiff device will be available if capabilities("aqua") || capabilities("tiff") if the defaults are unchanged.

See Also

.Platform

Examples

capabilities()

if(!capabilities("http/ftp"))
   warning("internal download.file() is not available")

## See also the examples for 'connections'.

[Package base version 3.1.0 Index]