cat {base}	R Documentation

Concatenate and Print

Description

Prints the arguments, coercing them if necessary to character mode first.

Usage

cat(... , file = "", sep = " ", fill = FALSE, labels = NULL,
    append = FALSE)

Arguments

... 	 R objects which are coerced to character strings, concatenated, and printed, with the remaining arguments controlling the output.

file 	 A connection, or a character string naming the file to print to. If "" (the default), cat prints to the standard output connection, the console unless redirected by sink.

sep 	 character string to insert between the objects to print.

fill 	 a logical or numeric controlling how the output is broken into successive lines. If FALSE (default), only newlines created explicitly by \n are printed. Otherwise, the output is broken into lines with print width equal to the option width if fill is TRUE, or the value of fill if this is numeric.

labels 	 character vector of labels for the lines printed. Ignored if fill is FALSE.

append 	 logical. Only used if the argument file is the name of file (and not a connection or "|cmd"). If TRUE output will be appended to file; otherwise, it will overwrite the contents of file.

Details

cat converts its arguments to character strings, concatenates them, separating them by the given sep= string, and then prints them.

No linefeeds are printed unless explicitly requested by \n or if generated by filling (if argument fill is TRUE or numeric.)

cat is useful for producing output in user-defined functions.

Value

None (invisible NULL).

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

print, format, and paste which concatenates into a string.

Examples

iter <- rpois(1, lambda=10)
## print an informative message
cat("iteration = ", iter <- iter + 1, "\n")

## 'fill' and label lines:
cat(paste(letters, 100* 1:26), fill = TRUE,
    labels = paste("{",1:10,"}:",sep=""))

[Package base version 2.1.0 Index]