copyright {base}	R Documentation

Copyrights of Files Used to Build R

Description

R is released under the ‘GNU Public License’: see license for details. The license describes your right to use R. Copyright is concerned with ownership of intellectual rights, and some of the software used has conditions that the copyright must be explicitly stated: see the ‘Details’ section. We are grateful to these people and other contributors (see contributors) for the ability to use their work.

Details

The file ‘R_HOME/COPYRIGHTS’ lists the copyrights in full detail.

[Package base version 3.6.1 Index]