libcurlVersion {base}	R Documentation

Report Version of libcurl

Description

Report version of libcurl in use.

Usage

libcurlVersion()

Value

A character string, with value the libcurl version in use or "" if none is. If libcurl is available, has attributes

ssl_version
A character string naming the SSL implementation and version, possibly "none". It is intended for the version of OpenSSL used, but not all implementations of libcurl use OpenSSL — for example OS X reports "SecureTranspart", its wrapper for SSL/TLS.

libssh_version
A character string naming the libssh version, which may or may not be available (it is used for e.g. scp and sftp protocols). Where present, something like "libssh2/1.5.0".

protocols
A character vector of the names of supported protocols, also known as ‘schemes’ when part of a URL.

See Also

extSoftVersion for versions of other third-party software.

curlGetHeaders, download.file and url for functions which (optionally) use libcurl.

http://curl.haxx.se/docs/sslcerts.html and http://curl.haxx.se/docs/ssl-compared.html for more details on SSL versions. Normally libcurl used with R uses SecureTransport on OS X, OpenSSL on Windows and GnuTLS, NSS or OpenSSL on Unix-alikes. (At the time of writing Debian-based Linuxen use GnuTLS, RedHat-based ones use NSS.)

Examples

libcurlVersion()

[Package base version 3.2.2 Index]