dataframeHelpers {base}	R Documentation

Data Frame Auxiliary Functions

Description

Internal auxiliary functions for use with data frames.

Usage


xpdrows.data.frame(x, old.rows, new.rows)

Arguments

x
object of class data.frame.

old.rows, new.rows
row names for old and new rows.

Details

xpdrows.data.frame is an auxiliary function which expands the rows of a data frame. It is used by the data frame methods of [<- and [[<- (which perform subscripted assignments on a data frame), and not intended to be called directly.

See Also

[.data.frame

[Package base version 3.6.1 Index]