.Device {base}	R Documentation

Lists of Open Graphics Devices

Description

A list of the names of the open graphics devices is stored in .Devices. The name of the active device is stored in .Device.

[Package base version 2.1.0 Index]