dontCheck {base}	R Documentation

Identity Function to Suppress Checking

Description

The dontCheck function is the same as identity, but is interpreted by R CMD check code analysis as a directive to suppress checking of x. Currently this is only used by checkFF(registration = TRUE) when checking the .NAME argument of foreign function calls.

Usage

dontCheck(x)

Arguments

x
an R object.

See Also

suppressForeignCheck which explains why that and dontCheck are undesirable and should be avoided if at all possible.

[Package base version 3.1.0 Index]