expression {base}	R Documentation

Unevaluated Expressions

Description

Creates or tests for objects of mode "expression".

Usage

expression(...)

is.expression(x)
as.expression(x, ...)

Arguments

... 	 valid R expressions.

x 	 an arbitrary R object.

Value

expression returns a vector of mode "expression" containing its arguments as unevaluated “calls”.
is.expression returns TRUE if expr is an expression object and FALSE otherwise.
as.expression attempts to coerce its argument into an expression object.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

call, eval, function. Further, text and legend for plotting math expressions.

Examples

length(ex1 <- expression(1+ 0:9))# 1
ex1
eval(ex1)# 1:10

length(ex3 <- expression(u,v, 1+ 0:9))# 3
mode(ex3 [3]) # expression
mode(ex3[[3]])# call
rm(ex3)

[Package base version 2.1.0 Index]