extSoftVersion {base}	R Documentation

Report Versions of Third-Party Software

Description

Report versions of (external) third-party software used.

Usage


extSoftVersion()

Details

The reports the versions of third-party software libraries in use. These are often external but might have been compiled into R when it was installed.

With dynamic linking, these are the versions of the libraries linked to in this session: with static linking, of those compiled in.

Value

A named character vector, currently with components

zlib
The version of zlib in use.

bzlib
The version of bzlib (from bzip2) in use.

xz
The version of liblzma (from xz) in use.

PCRE
The version of PCRE in use.

ICU
The version of ICU in use (if any, otherwise "").

TRE
The version of libtre in use.

iconv
The implementation and version of the iconv library in use (if known).

readline
The version of readline in use (if any, otherwise "").

BLAS
Name of the binary/executable file with the implementation of BLAS in use (if known, otherwise "").

Note that the values for bzlib and pcre normally contain a date as well as the version number, and that for tre includes several items separated by spaces, the version number being the second.

For iconv this will give the implementation as well as the version, for example "GNU libiconv 1.14", "glibc 2.18" or "win_iconv" (which has no version number).

The name of the binary/executable file for BLAS can be used as an indication of which implementation is in use. Typically, the R version of BLAS will appear as libR.so (libR.dylib), R or libRblas.so (libRblas.dylib), depending on how R was built. Note that libRblas.so (libRblas.dylib) may also be shown for an external BLAS implementation that had been copied, hard-linked or renamed by the system administrator. For an external BLAS, a shared object file will be given and its path/name may indicate the vendor/version. The detection does not work on Windows.

See Also

libcurlVersion for the version of libCurl.

La_version for the version of LAPACK in use.

La_library for binary/executable file with LAPACK in use.

grSoftVersion for third-party graphics software.

tclVersion for the version of Tcl/Tk.

pcre_config for PCRE configuration options.

Examples


extSoftVersion()
## the PCRE version
sub(" .*", "", extSoftVersion()["PCRE"])

[Package base version 3.6.1 Index]