file.choose {base}	R Documentation

Choose a File Interactively

Description

Choose a file interactively.

Usage

file.choose(new = FALSE)

Arguments

new
Logical: choose the style of dialog box presented to the user: at present only new = FALSE is used.

Value

A character vector of length one giving the file path.

See Also

list.files for non-interactive selection.

[Package base version 3.2.2 Index]