format.pval {base}	R Documentation

Format P Values

Description

format.pval is intended for formatting p-values.

Usage


format.pval(pv, digits = max(1, getOption("digits") - 2),
            eps = .Machine$double.eps, na.form = "NA", ...)

Arguments

pv
a numeric vector.

digits
how many significant digits are to be used.

eps
a numerical tolerance: see ‘Details’.

na.form
character representation of NAs.

...
further arguments to be passed to format such as nsmall.

Details

format.pval is mainly an auxiliary function for print.summary.lm etc., and does separate formatting for fixed, floating point and very small values; those less than eps are formatted as "< [eps]" (where ‘[eps]’ stands for format(eps, digits)).

Value

A character vector.

Examples


format.pval(c(stats::runif(5), pi^-100, NA))
format.pval(c(0.1, 0.0001, 1e-27))

[Package base version 3.6.1 Index]