Sys.getenv {base}	R Documentation

Get Environment Variables

Description

Sys.getenv obtains the values of the environment variables named by x.

Usage

Sys.getenv(x)

Arguments

x 	 a character vector, or missing

Value

A vector of the same length as x, with the variable names as its names attribute. Each element holds the value of the environment variable named by the corresponding component of x (or "" if no environment variable with that name was found).
On most platforms Sys.getenv() will return a named vector giving the values of all the environment variables.

See Also

Sys.putenv, Sys.getlocale for the locale “environment”, getwd for the working directory.

Examples

## whether HOST is set will be shell-dependent e.g. Solaris' csh does not.
Sys.getenv(c("R_HOME", "R_PAPERSIZE", "R_PRINTCMD", "HOST"))

str(s <- Sys.getenv()) # all settings (rather do not print)

## Language and Locale settings -- but rather use Sys.getlocale()
s[grep("^L(C|ANG)", names(s))]

[Package base version 2.1.0 Index]