gl {base}	R Documentation

Generate Factor Levels

Description

Generate factors by specifying the pattern of their levels.

Usage


gl(n, k, length = n*k, labels = seq_len(n), ordered = FALSE)

Arguments

n
an integer giving the number of levels.

k
an integer giving the number of replications.

length
an integer giving the length of the result.

labels
an optional vector of labels for the resulting factor levels.

ordered
a logical indicating whether the result should be ordered or not.

Value

The result has levels from 1 to n with each value replicated in groups of length k out to a total length of length.

gl is modelled on the GLIM function of the same name.

See Also

The underlying factor().

Examples


## First control, then treatment:
gl(2, 8, labels = c("Control", "Treat"))
## 20 alternating 1s and 2s
gl(2, 1, 20)
## alternating pairs of 1s and 2s
gl(2, 2, 20)

[Package base version 3.6.1 Index]