is.language {base}	R Documentation

Is an Object a Language Object?

Description

is.language returns TRUE if x is a variable name, a call, or an expression.

Usage


is.language(x)

Arguments

x
object to be tested.

Note

A name is also known as ‘symbol’, from its type (typeof), see is.symbol.

If typeof(x) == "language", then is.language(x) is always true, but the reverse does not hold as expressions or names y also fulfill is.language(y), see the examples.

This is a primitive function.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Examples


ll <- list(a = expression(x^2 - 2*x + 1), b = as.name("Jim"),
           c = as.expression(exp(1)), d = call("sin", pi))
sapply(ll, typeof)
sapply(ll, mode)
stopifnot(sapply(ll, is.language))

[Package base version 3.6.1 Index]