is.object {base}	R Documentation

Is an Object “internally classed”?

Description

A function rather for internal use. It returns TRUE if the object x has the R internal OBJECT attribute set, and FALSE otherwise.

Usage

is.object(x)

Arguments

x 	 object to be tested.

Details

is.object is generic: you can write methods to handle specific classes of objects, see InternalMethods.

See Also

class, and methods.

Examples

is.object(1) # FALSE
is.object(as.factor(1:3)) # TRUE

[Package base version 2.1.0 Index]