kappa {base}	R Documentation

Estimate the Condition Number

Description

An estimate of the condition number of a matrix or of the R matrix of a QR decomposition, perhaps of a linear fit. The condition number is defined as the ratio of the largest to the smallest non-zero singular value of the matrix.

Usage

kappa(z, ...)
## S3 method for class 'lm':
kappa(z, ...)
## Default S3 method:
kappa(z, exact = FALSE, ...)
## S3 method for class 'qr':
kappa(z, ...) 

kappa.tri(z, exact = FALSE, ...)

Arguments

z 	 A matrix or a the result of qr or a fit from a class inheriting from "lm".

exact 	 logical. Should the result be exact?

... 	 further arguments passed to or from other methods.

Details

If exact = FALSE (the default) the condition number is estimated by a cheap approximation. Following S, this uses the LINPACK routine ‘dtrco.f’. However, in R (or S) the exact calculation is also likely to be quick enough.

kappa.tri is an internal function called by kappa.qr.

Value

The condition number, kappa, or an approximation if exact = FALSE.

Author(s)

The design was inspired by (but differs considerably from) the S function of the same name described in Chambers (1992).

References

Chambers, J. M. (1992) Linear models. Chapter 4 of Statistical Models in S eds J. M. Chambers and T. J. Hastie, Wadsworth & Brooks/Cole.

See Also

svd for the singular value decomposition and qr for the QR one.

Examples

kappa(x1 <- cbind(1,1:10))# 15.71
kappa(x1, exact = TRUE)        # 13.68
kappa(x2 <- cbind(x1,2:11))# high! [x2 is singular!]

hilbert <- function(n) { i <- 1:n; 1 / outer(i - 1, i, "+") }
sv9 <- svd(h9 <- hilbert(9))$ d
kappa(h9)# pretty high!
kappa(h9, exact = TRUE) == max(sv9) / min(sv9)
kappa(h9, exact = TRUE) / kappa(h9) # .677 (i.e., rel.error = 32%)

[Package base version 2.1.0 Index]