l10n_info {base}	R Documentation

Localization Information

Description

Report on localization information.

Usage


l10n_info()

Details

Common codepages are 1252 (Western European), 1250 (Central European), 1251 (Cyrillic), 1253 (Greek), 1254 (Turkish), 1255 (Hebrew), 1256 (Arabic), 1257 (Baltic), 1258 (Vietnamese), 874 (Thai), 932 (Japanese), 936 (Simplified Chinese), 949 (Korean) and 950 (Traditional Chinese). R does not allow the C locale, and uses 1252 as the default codepage.

Value

A list with three logical components:

MBCS
If a multi-byte character set in use?

UTF-8
Is this a UTF-8 locale?

Latin-1
Is this a Latin-1 locale?

codepage
integer: the Windows codepage corresponding to the locale R is using (and not necessarily that Windows is using).

See Also

Sys.getlocale, localeconv

Examples


l10n_info()

[Package base version 3.6.1 Index]