lapply {base}	R Documentation

Apply a Function over a List or Vector

Description

lapply returns a list of the same length as X. Each element of which is the result of applying FUN to the corresponding element of X.

sapply is a “user-friendly” version of lapply also accepting vectors as X, and returning a vector or matrix with dimnames if appropriate.

replicate is a wrapper for the common use of sapply for repeated evaluation of an expression (which will usually involve random number generation).

Usage

lapply(X, FUN, ...)
sapply(X, FUN, ..., simplify = TRUE, USE.NAMES = TRUE)

replicate(n, expr, simplify = TRUE) 

Arguments

X 	 list or vector to be used.

FUN 	 the function to be applied. In the case of functions like +, %*%, etc., the function name must be quoted.

... 	 optional arguments to FUN.

simplify 	 logical; should the result be simplified to a vector or matrix if possible?

USE.NAMES 	 logical; if TRUE and if X is character, use X as names for the result unless it had names already.

n 	 Number of replications.

expr 	 Expression to evaluate repeatedly.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

apply, tapply.

Examples

x <- list(a = 1:10, beta = exp(-3:3), logic = c(TRUE,FALSE,FALSE,TRUE))
# compute the list mean for each list element
lapply(x,mean)
# median and quartiles for each list element
lapply(x, quantile, probs = 1:3/4)
sapply(x, quantile)
i39 <- sapply(3:9, seq) # list of vectors
sapply(i39, fivenum)

hist(replicate(100, mean(rexp(10))))

[Package base version 2.1.0 Index]