license {base}	R Documentation

The R License Terms

Description

The license terms under which R is distributed.

Usage


license()
licence()

Details

R is distributed under the terms of the GNU GENERAL PUBLIC LICENSE, either Version 2, June 1991 or Version 3, June 2007. A copy of the version 2 license is in file ‘R_HOME/doc/COPYING’ and can be viewed by RShowDoc("COPYING"). Version 3 of the license can be displayed by RShowDoc("GPL-3").

A small number of files (some of the API header files) are distributed under the LESSER GNU GENERAL PUBLIC LICENSE, version 2.1 or later. A copy of this license is in file ‘$R_SHARE_DIR/licenses/LGPL-2.1’ and can be viewed by RShowDoc("LGPL-2.1"). Version 3 of the license can be displayed by RShowDoc("LGPL-3").

[Package base version 3.6.1 Index]