manglePackageName {base}	R Documentation

Mangle the Package Name

Description

This function takes the package name and the package version number and pastes them together with a separating underscore.

Usage

manglePackageName(pkgName, pkgVersion)

Arguments

pkgName 	 The package name, as a character string.

pkgVersion 	 The package version, as a character string.

Value

A character string with the two inputs pasted together.

Examples

  manglePackageName("foo", "1.2.3")

[Package base version 2.7.2 Index]