margin.table {base}	R Documentation

Compute table margin

Description

For a contingency table in array form, compute the sum of table entries for a given index.

Usage


margin.table(x, margin = NULL)

Arguments

x
an array

margin
index number (1 for rows, etc.)

Details

This is really just apply(x, margin, sum) packaged up for newbies, except that if margin has length zero you get sum(x).

Value

The relevant marginal table. The class of x is copied to the output table, except in the summation case.

Author(s)

Peter Dalgaard

See Also

prop.table and addmargins.

Examples


m <- matrix(1:4, 2)
margin.table(m, 1)
margin.table(m, 2)

[Package base version 3.6.1 Index]