memory.profile {base}	R Documentation

Profile the Usage of Cons Cells

Description

Lists the usage of the cons cells by SEXPREC type.

Usage

memory.profile()

Details

The current types and their uses are listed in the include file ‘Rinternals.h’. There will be blanks in the list corresponding to types that are no longer in use (types 11 and 12 at the time of writing). Also FUNSXP is not included.

Value

A vector of counts, named by the types.

See Also

gc for the overall usage of cons cells.

Examples

memory.profile()

[Package base version 2.1.0 Index]