message {base}	R Documentation

Diagnostic Messages

Description

Generate a diagnostic message from its arguments.

Usage

message(..., domain = NULL)
suppressMessages(expr)

Arguments

... 	 character vectors (which are pasted together with no separator), a condition object, or NULL.

domain 	 see gettext. If NA, messages will not be translated.

expr 	 expression to evaluate.

Details

message is used for generating “simple” diagnostic messages which are neither warnings nor errors, but nevertheless represented as conditions.

While the message is being processed, a muffleMessage restart is available.

suppressMessages evaluates its expression in a context that ignores all “simple” diagnostic messages.

See Also

warning and stop for generating warnings and errors; conditions for condition handling and recovery.

gettext for the mechanisms for the automated translation of text.

Examples

message("ABC", "DEF")
suppressMessages(message("ABC"))

[Package base version 2.1.0 Index]