NotYet {base}	R Documentation

Not Yet Implemented Functions and Unused Arguments

Description

In order to pinpoint missing functionality, the R core team uses these functions for missing R functions and not yet used arguments of existing R functions (which are typically there for compatibility purposes).

You are very welcome to contribute your code ...

Usage


.NotYetImplemented()
.NotYetUsed(arg, error = TRUE)

Arguments

arg
an argument of a function that is not yet used.

error
a logical. If TRUE, an error is signalled; if FALSE; only a warning is given.

See Also

the contrary, Deprecated and Defunct for outdated code.

Examples


require(graphics)
barplot(1:5, inside = TRUE) # 'inside' is not yet used

[Package base version 3.6.1 Index]