ns-lowlev {base}	R Documentation

Low Level Name Space Support Functions

Description

Low level name space support functions.

Usage

loadingNamespaceInfo()
saveNamespaceImage(package, rdafile, lib.loc = NULL,
                   keep.source = getOption("keep.source.pkgs"),
                   compress = TRUE)

Arguments

package 	 string naming the package/name space to load.

rdafile 	 name of file to save to.

lib.loc 	 character vector specifying library search path.

keep.source 	 logical specifying whether to retain source.

compress 	 option passed to save.

Details

The functions loadNamespace and attachNamespace are usually called implicitly when library is used to load a name space and any imports needed. However it may be useful to call these functions directly at times.

loadNamespace loads the specified name space and registers it in an internal data base. A request to load a name space that is already loaded has no effect. The arguments have the same meaning as the corresponding arguments to library. After loading, loadNamespace looks for a hook function named .onLoad as an internal variable in the name space (it should not be exported). This function is called with the same arguments as .First.lib. Partial loading is used to support installation with the --save option.

loadNamespace does not attach the name space it loads to the search path. attachNamespace can be used to attach a frame containing the exported values of a name space to the search path. The hook function .onAttach is run after the name space exports are attached.

loadedNamespaces returns a character vector of the names of the loaded name spaces.

unloadNamespace can be used to force a name space to be unloaded. An error is signaled if the name space is imported by other loaded name spaces. If defined, a hook function .onUnload is run before removing the name space from the internal registry. unloadNamespace will first detach a package of the same name if one is on the path, thereby running a .Last.lib function in the package if one is exported.

loadingNamespaceInfo returns a list of the arguments that would be passed to .onLoad when a name space is being loaded. An error is signaled of a name space is not currently being loaded.

saveNamespaceImage is used to save name space images for packages installed with --save.

Author(s)

Luke Tierney

[Package base version 2.9.2 Index]