ns-reflect.Rd {base}	R Documentation

Namespace Reflection Support

Description

Internal functions to support reflection on namespace objects.

Usage


getExportedValue(ns, name)
getNamespace(name)
getNamespaceExports(ns)
getNamespaceImports(ns)
getNamespaceName(ns)
getNamespaceUsers(ns)
getNamespaceVersion(ns)

Arguments

ns
string or namespace object.

name
string or name.

Details

getExportedValue returns the value of the exported variable name in namespace ns.

getNamespace returns the environment representing the name space name. The namespace is loaded if necessary.

getNamespaceExports returns a character vector of the names exported by ns.

getNamespaceImports returns a representation of the imports used by namespace ns. This representation is experimental and subject to change.

getNamespaceName and getNamespaceVersion return the name and version of the namespace ns.

getNamespaceUsers returns a character vector of the names of the namespaces that import namespace ns.

Author(s)

Luke Tierney

See Also

loadNamespace for more about namespaces.

[Package base version 3.6.1 Index]