numeric {base}	R Documentation

Numeric Vectors

Description

Creates or tests for objects of type "numeric".

Usage

numeric(length = 0)
as.numeric(x, ...)
is.numeric(x)

Arguments

length 	 desired length.

x 	 object to be coerced or tested.

... 	 further arguments passed to or from other methods.

Details

as.numeric is a generic function, but methods must be written for as.double, which it calls.

is.numeric is generic: you can write methods to handle specific classes of objects, see InternalMethods.

Note that factors are false for is.numeric since there is a "factor" method.

Value

numeric creates a real vector of the specified length. The elements of the vector are all equal to 0.
as.numeric attempts to coerce its argument to numeric type (either integer or real). as.numeric for factors yields the codes underlying the factor levels, not the numeric representation of the labels.
is.numeric returns TRUE if its argument is of type numeric or type integer and FALSE otherwise.

Note

R has no single precision data type. All real numbers are stored in double precision format.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Examples

as.numeric(c("-.1"," 2.7 ","B")) # (-0.1, 2.7, NA)  +  warning
as.numeric(factor(5:10))

[Package base version 2.1.0 Index]