path.expand {base}	R Documentation

Expand File Paths

Description

Expand a path name, for example by replacing a leading tilde by the user's home directory (if defined on that platform).

Usage

path.expand(path)

Arguments

path
character vector containing one or more path names.

Details

On some Unix builds of R, a leading ~user will expand to the home directory of user, but not on Unix versions without readline installed, nor if R is invoked with --no-readline.

In an interactive session capabilities("cledit") will report if readline is available.

See Also

basename, normalizePath.

Examples

path.expand("~/foo")

[Package base version 2.16.0 Index]