print.data.frame {base}	R Documentation

Printing Data Frames

Description

Print a data frame.

Usage


## S3 method for class 'data.frame'
print(x, ..., digits = NULL,
      quote = FALSE, right = TRUE, row.names = TRUE, max = NULL)

Arguments

x
object of class data.frame.

...
optional arguments to print or plot methods.

digits
the minimum number of significant digits to be used: see print.default.

quote
logical, indicating whether or not entries should be printed with surrounding quotes.

right
logical, indicating whether or not strings should be right-aligned. The default is right-alignment.

row.names
logical (or character vector), indicating whether (or what) row names should be printed.

max
numeric or NULL, specifying the maximal number of entries to be printed. By default, when NULL, getOption("max.print") used.

Details

This calls format which formats the data frame column-by-column, then converts to a character matrix and dispatches to the print method for matrices.

When quote = TRUE only the entries are quoted not the row names nor the column names.

See Also

data.frame.

Examples


(dd <- data.frame(x = 1:8, f = gl(2,4), ch = I(letters[1:8])))
     # print() with defaults
print(dd, quote = TRUE, row.names = FALSE)
     # suppresses row.names and quotes all entries

[Package base version 3.6.1 Index]