print.default {base}	R Documentation

Default Printing

Description

print.default is the default method of the generic print function which prints its argument.

Usage

## Default S3 method:
print(x, digits = NULL, quote = TRUE, na.print = NULL,
      print.gap = NULL, right = FALSE, ...)

Arguments

x 	 the object to be printed.

digits 	 a non-null value for digits specifies the minimum number of significant digits to be printed in values. If digits is NULL, the value of digits set by options is used.

quote 	 logical, indicating whether or not strings (characters) should be printed with surrounding quotes.

na.print 	 a character string which is used to indicate NA values in printed output, or NULL (see Details)

print.gap 	 a non-negative integer <= 1024, giving the spacing between adjacent “columns” in printed vectors, matrices and arrays, or NULL meaning 1.

right 	 logical, indicating whether or not strings should be right-aligned. The default is left-alignment.

... 	 further arguments to be passed to or from other methods. They are ignored in this function.

Details

The default for printing NAs is to print NA (without quotes) unless this is a character NA and quote = FALSE, when <NA> is printed.

The same number of decimal places is used throughout a vector, This means that digits specifies the minimum number of significant digits to be used, and that at least one entry will be printed with that minimum number.

Attributes are printed respecting their class(es), using the values of digits to print.default, but using the default values (for the methods called) of the other arguments.

When the methods package is attached, print will call show for R objects with formal classes if called with no optional arguments.

If a non-printable character is encountered during output, it is represented as one of the ANSI escape sequences (\a, \b, \f, \n, \r, \t, \v and \0), or failing that as a 3-digit octal code: for example the UK currency pound in the C locale (if implemented correctly) is printed as \243. Which characters are non-printable depends on the locale. (Because some versions of Windows get this wrong, all 8-bit characters are regarded as printable on Windows.)

Unicode and other multi-byte locales

In a Unicode (UTF-8) locale, characters 0x00 to 0x1F and 0x7f (the ASCII non-printing characters) are printed in the same way, via ANSI escape sequences or 3-digit octal escapes. Multi-byte non-printing characters are printed with as an escape sequence of the form \uxxxx or \Uxxxxxxxx (in hexadecimal).

It is possible to have a character string in an object that is not valid UTF-8. If a byte is encountered that is not part of an encoded Unicode character it is printed in hex in the form <xx> and the next character is tried.

Note

print.matrix is currently identical to print.default, but was prior to 1.7.0 did not print attributes and did not have a digits argument. It is provided only because some packages call it explicitly.

See Also

The generic print, options. The "noquote" class and print method.

encodeString.

Examples

pi
print(pi, digits = 16)
LETTERS[1:16]
print(LETTERS, quote = FALSE)

[Package base version 2.1.0 Index]