prmatrix {base}	R Documentation

Print Matrices, Old-style

Description

An earlier method for printing matrices, provided for S compatibility.

Usage


prmatrix(x, rowlab =, collab =,
         quote = TRUE, right = FALSE, na.print = NULL, ...)

Arguments

x
numeric or character matrix.

rowlab, collab
(optional) character vectors giving row or column names respectively. By default, these are taken from dimnames(x).

quote
logical; if TRUE and x is of mode "character", quotes (") are used.

right
if TRUE and x is of mode "character", the output columns are right-justified.

na.print
how NAs are printed. If this is non-null, its value is used to represent NA.

...
arguments for print methods.

Details

prmatrix is an earlier form of print.matrix, and is very similar to the S function of the same name.

Value

Invisibly returns its argument, x.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

print.default, and other print methods.

Examples


prmatrix(m6 <- diag(6), rowlab = rep("", 6), collab = rep("", 6))

chm <- matrix(scan(system.file("help", "AnIndex", package = "splines"),
                   what = ""), , 2, byrow = TRUE)
chm  # uses print.matrix()
prmatrix(chm, collab = paste("Column", 1:3), right = TRUE, quote = FALSE)

[Package base version 3.6.1 Index]