prod {base}	R Documentation

Product of Vector Elements

Description

prod returns the product of all the values present in its arguments.

Usage

prod(..., na.rm = FALSE)

Arguments

... 	 numeric vectors.

na.rm 	 logical. Should missing values be removed?

Details

If na.rm is FALSE an NA value in any of the arguments will cause a value of NA to be returned, otherwise NA values are ignored.

This is a generic function: methods can be defined for it directly or via the Summary group generic.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

sum, cumprod, cumsum.

Examples

print(prod(1:7)) == print(gamma(8))

[Package base version 2.1.0 Index]