readRenviron {base}	R Documentation

Set Environment Variables from a File

Description

Read as file such as ‘.Renviron’ or ‘Renviron.site’ in the format described in the help for Startup, and set environment variables as defined in the file.

Usage


readRenviron(path)

Arguments

path
A length-one character vector giving the path to the file. Tilde-expansion is performed where supported.

Value

Scalar logical indicating if the file was read successfully. Returned invisibly. If the file cannot be opened for reading, a warning is given.

See Also

Startup for the file format.

Examples

## Not run: 
## re-read a startup file (or read it in a vanilla session)
readRenviron("~/.Renviron")

## End(Not run)

[Package base version 3.6.1 Index]