real {base}	R Documentation

Real Vectors

Description

real creates a double precision vector of the specified length. Each element of the vector is equal to 0.

as.real attempts to coerce its argument to be of real type.

is.real returns TRUE or FALSE depending on whether its argument is of real type or not.

Usage

real(length = 0)
as.real(x, ...)
is.real(x)

Arguments

length 	 desired length.

x 	 object to be coerced or tested.

... 	 further arguments passed to or from other methods.

Note

R has no single precision data type. All real numbers are stored in double precision format.

[Package base version 2.1.0 Index]