shell.exec {base}	R Documentation

Open a File using Windows File Associations

Description

Opens the specified file using the application specified in the Windows file associations.

Usage

shell.exec(file)

Arguments

file 	 file to be opened.

Details

It is likely that file needs to be a complete path, e.g. ‘"c:/R/results.html"’ or ‘"c:R\results.html"’ as most applications interpret files relative to their working directory.

Value

No value.

Author(s)

B. D. Ripley

See Also

system, shell

[Package base version 2.1.0 Index]