slotOp {base}	R Documentation

Extract Slots

Description

Extract tbe contents of a slot in a object with a formal class structure.

Usage

object@name

Arguments

object 	 An object from a formally defined class.

name 	 The character-string name of the slot.

Details

These operators support the formal classes of package methods. See slot for further details. Currently there is no checking that the object is an instance of a class.

See Also

Extract, slot

[Package base version 2.1.0 Index]