strtrim {base}	R Documentation

Trim Character Strings to Specified Widths

Description

Trim character strings to specified display widths.

Usage

strtrim(x, width)

Arguments

x 	 a character vector, or an object which can be coerced to a character vector.

width 	 Positive integer values: recycled to the length of x.

Details

‘Width’ is interpreted as the display width in a monospaced font. What happens with non-printable characters (such as backspace, tab) is implementation-dependent and may depend on the locale (e.g. they may be included in the count or they may be omitted).

Using this function rather than substr is important when there might be double-width characters in character vectors

Value

A character vector of the same length as x.

Examples

strtrim(c("abcdef", "abcdef", "abcdef"), c(1,5,10))

[Package base version 2.1.0 Index]