structure {base}	R Documentation

Attribute Specification

Description

structure returns the given object with its attributes set.

Usage

structure(.Data, ...)

Arguments

.Data 	 an object which will have various attributes attached to it.

... 	 attributes, specified in tag=value form, which will be attached to data.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Examples

structure(1:6, dim = 2:3)

[Package base version 2.1.0 Index]