unique {base}	R Documentation

Extract Unique Elements

Description

unique returns a vector, data frame or array like x but with duplicate elements removed.

Usage

unique(x, incomparables = FALSE, ...)

## S3 method for class 'array':
unique(x, incomparables = FALSE, MARGIN = 1, ...)

Arguments

x 	 a vector or a data frame or an array or NULL.

incomparables 	 a vector of values that cannot be compared. Currently, FALSE is the only possible value, meaning that all values can be compared.

... 	 arguments for particular methods.

MARGIN 	 the array margin to be held fixed: a single integer.

Details

This is a generic function with methods for vectors, data frames and arrays (including matrices).

The array method calculates for each element of the dimension specified by MARGIN if the remaining dimensions are identical to those for an earlier element (in row-major order). This would most commonly be used to find unique rows (the default) or columns (with MARGIN = 2).

Value

An object of the same type of x. but if an element is equal to one with a smaller index, it is removed. Dimensions of arrays are not dropped.

Warning

Using this for lists is potentially slow, especially if the elements are not atomic vectors (see vector) or differ only in their attributes. In the worst case it is O(n^2).

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

duplicated which gives the indices of duplicated elements.

Examples

unique(c(3:5, 11:8, 8 + 0:5))
length(unique(sample(100, 100, replace=TRUE)))
## approximately 100(1 - 1/e) = 63.21

unique(iris)

[Package base version 2.1.0 Index]