unname {base}	R Documentation

Remove ‘names’ or ‘dimnames’

Description

Remove the names or dimnames attribute of an R object.

Usage

unname(obj, force = FALSE)

Arguments

obj 	 an R object.

force 	 logical; if true, the dimnames are even removed from data.frames. This argument is currently experimental and hence might change!

Value

Object as obj but without names or dimnames.

Examples

require(graphics); require(stats)

## Answering a question on R-help (14 Oct 1999):
col3 <- 750+ 100*rt(1500, df = 3)
breaks <- factor(cut(col3,breaks=360+5*(0:155)))
z <- table(breaks)
z[1:5] # The names are larger than the data ...
barplot(unname(z), axes= FALSE)

[Package base version 2.9.2 Index]