write {base}	R Documentation

Write Data to a File

Description

The data (usually a matrix) x are written to file file. If x is a two-dimensional matrix you need to transpose it to get the columns in file the same as those in the internal representation.

Usage

write(x, file = "data",
      ncolumns = if(is.character(x)) 1 else 5,
      append = FALSE)

Arguments

x 	 the data to be written out.

file 	 A connection, or a character string naming the file to write to. If "", print to the standard output connection.

ncolumns 	 the number of columns to write the data in.

append 	 if TRUE the data x is appended to file file.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

save for writing any R objects, write.table for data frames, and scan for reading data.

Examples

# create a 2 by 5 matrix
x <- matrix(1:10,ncol=5)

# the file data contains x, two rows, five cols
# 1 3 5 6 9 will form the first row
write(t(x))

# the file data now contains the data in x,
# two rows, five cols but the first row is 1 2 3 4 5
write(x)
unlink("data") # tidy up

[Package base version 2.1.0 Index]