InternalMethods {base}	R Documentation

Internal Generic Functions

Description

Many R-internal functions are generic and allow methods to be written for.

Details

The following builtin functions are generic as well, i.e., you can write methods for them:

[, [[, $, [<-, [[<-, $<-,

length, length<-,

dimnames<-, dimnames, dim<-, dim

c, unlist,

as.character, as.vector, is.array, is.atomic, is.call, is.character, is.complex, is.double, is.environment, is.function, is.integer, is.language, is.logical, is.list, is.matrix, is.na, is.nan is.null, is.numeric, is.object, is.pairlist, is.recursive, is.single, is.symbol.

See Also

methods for the methods of non-Internal generic functions.

[Package base version 2.1.0 Index]