Constants {base}	R Documentation

Built-in Constants

Description

Constants built into R.

Usage


LETTERS
letters
month.abb
month.name
pi

Details

R has a small number of built-in constants.

The following constants are available:

LETTERS: the 26 upper-case letters of the Roman alphabet;

letters: the 26 lower-case letters of the Roman alphabet;

month.abb: the three-letter abbreviations for the English month names;

month.name: the English names for the months of the year;

pi: the ratio of the circumference of a circle to its diameter.

These are implemented as variables in the base namespace taking appropriate values.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

data, DateTimeClasses.

Quotes for the parsing of character constants, NumericConstants for numeric constants.

Examples


## John Machin (ca 1706) computed pi to over 100 decimal places
## using the Taylor series expansion of the second term of
pi - 4*(4*atan(1/5) - atan(1/239))

## months in English
month.name
## months in your current locale
format(ISOdate(2000, 1:12, 1), "%B")
format(ISOdate(2000, 1:12, 1), "%b")

[Package base version 3.6.1 Index]