Cstack_info {base}	R Documentation

Report Information on C Stack Size and Usage

Description

Report information on the C stack size and usage (if available).

Usage

Cstack_info()

Details

On most platforms, C stack information is recorded when R is initialized and used for stack-checking. If this information is unavailable, the size will be returned as NA, and stack-checking is not performed.

The information on the stack base address is thought to be accurate on Windows, Linux and FreeBSD (including Mac OS X), but a heuristic is used on other platforms. Because this might be slightly inaccurate, the current usage could be estimated as negative. (The heuristic is not used on embedded uses of R on platforms where the stack base is not thought to be accurate.)

Value

An integer vector. This has named elements

size 	 The size of the stack (in bytes), or NA if unknown.

current 	 The estimated current usage (in bytes), possibly NA.

direction 	 1 (stack grows down, the usual case) or -1 (stack grows up).

eval_depth 	 The current evaluation depth (including two calls for the call to Cstack_info).

Examples

Cstack_info()

[Package base version 2.9.2 Index]