Dates {base}	R Documentation

Date Class

Description

Description of the class "Date" representing calendar dates.

Usage


## S3 method for class 'Date'
summary(object, digits = 12, ...)

## S3 method for class 'Date'
print(x, max = NULL, ...)



Arguments

object, x
a Date object to be summarized or printed.

digits
number of significant digits for the computations.

max
numeric or NULL, specifying the maximal number of entries to be printed. By default, when NULL, getOption("max.print") used.

...
further arguments to be passed from or to other methods.

Details

Dates are represented as the number of days since 1970-01-01, with negative values for earlier dates. They are always printed following the rules of the current Gregorian calendar, even though that calendar was not in use long ago (it was adopted in 1752 in Great Britain and its colonies).

It is intended that the date should be an integer, but this is not enforced in the internal representation. Fractional days will be ignored when printing. It is possible to produce fractional days via the mean method or by adding or subtracting (see Ops.Date).

From the many methods, see methods(class = "Date"), a few are documented separately, see below.

See Also

Sys.Date for the current date.

weekdays for convenience extraction functions.

Methods with extra arguments and documentation:

Ops.Date

for operators on "Date" objects.

format.Date

for conversion to and from character strings.

axis.Date

and hist.Date for plotting.

seq.Date

, cut.Date, and round.Date for utility operations.

DateTimeClasses for date-time classes.

Examples


(today <- Sys.Date())
format(today, "%d %b %Y")  # with month as a word
(tenweeks <- seq(today, length.out=10, by="1 week")) # next ten weeks
weekdays(today)
months(tenweeks)

(Dls <- as.Date(.leap.seconds))

##  length(<Date>) <- n   now works
ls <- Dls; length(ls) <- 12
l2 <- Dls; length(l2) <- 5 + length(Dls)
stopifnot(exprs = {
  ## length(.) <- * is compatible to subsetting/indexing:
  identical(ls, Dls[seq_along(ls)])
  identical(l2, Dls[seq_along(l2)])
  ## has filled with NA's
  is.na(l2[(length(Dls)+1):length(l2)])
})

[Package base version 3.6.1 Index]