Deprecated {base}	R Documentation

Marking Objects as Deprecated

Description

When an object is about removed from R it is first deprecated and should include a call to .Deprecated.

Usage

.Deprecated(new, package=NULL, msg)

Arguments

new 	 character string: A suggestion for a replacement function.

package 	 character string: The package to be used when suggesting where the deprecated function might be listed.

msg 	 character string: A message to be printed, if missing a default message is used.

Details

.Deprecated("<new name>") is called from deprecated functions. The original help page for these functions is often available at help("oldName-deprecated") (note the quotes). Functions should be listed in help("pkg-deprecated") for an appropriate pkg, including base.

See Also

Defunct

base-deprecated and so on which list the deprecated functions in the packages.

[Package base version 2.9.2 Index]