EnvVar {base}	R Documentation

Environment Variables

Description

Details of some of the environment variables which affect an R session.

Details

It is impossible to list all the environment variables which can affect an R session: some affect the OS system functions which R uses, and others will affect add-on packages. But here are notes on some of the more important ones. Those that set the defaults for options are consulted only at startup (as are some of the others).

HOME:

The user's ‘home’ directory.

LANGUAGE:

Optional. The language(s) to be used for message translations. This is consulted when needed.

LC_ALL:

(etc) Optional. Use to set various aspects of the locale – see Sys.getlocale. Consulted at startup.

MAKEINDEX:

The path to makeindex. If unset to a value determined when R was built. Used by the emulation mode of texi2dvi and texi2pdf.

R_BATCH:

Optional – set in a batch session, that is one started by R CMD BATCH. Most often set to "", so test by something like !is.na(Sys.getenv("R_BATCH", NA)).

R_BROWSER:

The path to the default browser. Used to set the default value of options("browser").

R_COMPLETION:

Optional. If set to FALSE, command-line completion is not used. (Not used by the macOS GUI.)

R_DEFAULT_PACKAGES:

A comma-separated list of packages which are to be attached in every session. See options.

R_DOC_DIR:

The location of the R ‘doc’ directory. Set by R.

R_ENVIRON:

Optional. The path to the site environment file: see Startup. Consulted at startup.

R_GSCMD:

Optional. The path to Ghostscript, used by dev2bitmap, bitmap and embedFonts. Consulted when those functions are invoked. Since it will be treated as if passed to system, spaces and shell metacharacters should be escaped.

R_HISTFILE:

Optional. The path of the history file: see Startup. Consulted at startup and when the history is saved.

R_HISTSIZE:

Optional. The maximum size of the history file, in lines. Exactly how this is used depends on the interface.

On Unix-alikes,

for the readline command-line interface it takes effect when the history is saved (by savehistory or at the end of a session).

On Windows,

for Rgui it controls the number of lines saved to the history file: the size of the history used in the session is controlled by the console customization: see Rconsole.

R_HOME:

The top-level directory of the R installation: see R.home. Set by R.

R_INCLUDE_DIR:

The location of the R ‘include’ directory. Set by R.

R_LIBS:

Optional. Used for initial setting of .libPaths.

R_LIBS_SITE:

Optional. Used for initial setting of .libPaths.

R_LIBS_USER:

Optional. Used for initial setting of .libPaths.

R_PAPERSIZE:

Optional. Used to set the default for options("papersize"), e.g. used by pdf and postscript.

R_PCRE_JIT_STACK_MAXSIZE:

Optional. Consulted when PCRE's JIT pattern compiler is first used. See grep.

R_PDFVIEWER:

The path to the default PDF viewer. Used by R CMD Rd2pdf.

R_PLATFORM:

The platform – a string of the form cpu-vendor-os, see R.Version.

R_PROFILE:

Optional. The path to the site profile file: see Startup. Consulted at startup.

R_RD4PDF:

Options for pdflatex processing of Rd files. Used by R CMD Rd2pdf.

R_SHARE_DIR:

The location of the R ‘share’ directory. Set by R.

R_TEXI2DVICMD:

The path to texi2dvi. Defaults to the value of TEXI2DVI, and if that is unset to a value determined when R was built.

Only on Unix-alikes:
Consulted at startup to set the default for options("texi2dvi"), used by texi2dvi and texi2pdf in package tools.

R_UNZIPCMD:

The path to unzip. Sets the initial value for options("unzip") on a Unix-alike when namespace utils is loaded.

R_ZIPCMD:

The path to zip. Used by zip and by R CMD INSTALL --build on Windows.

TMPDIR, TMP, TEMP:

Consulted (in that order) when setting the temporary directory for the session: see tempdir. TMPDIR is also used by some of the utilities see the help for build.

TZ:

Optional. The current time zone. See Sys.timezone for the system-specific formats. Consulted as needed.

no_proxy, http_proxy, ftp_proxy:

(and more). Optional. Settings for download.file: see its help for further details.

Unix-specific

Some variables set on Unix-alikes, and not (in general) on Windows.

DISPLAY:

Optional: used by X11, Tk (in package tcltk), the data editor and various packages.

EDITOR:

The path to the default editor: sets the default for options("editor") when namespace utils is loaded.

PAGER:

The path to the pager with the default setting of options("pager"). The default value is chosen at configuration, usually as the path to less.

R_PRINTCMD:

Sets the default for options("printcmd"), which sets the default print command to be used by postscript.

R_SUPPORT_OLD_TARS

logical. Sets the default for the support_old_tars argument of untar. Should be set to TRUE if an old system tar command is used which does not support either xz compression or automagically detecting compression type.

Windows-specific

Some Windows-specific variables are

GSC:

Optional: the path to Ghostscript, used if R_GSCMD is not set.

R_USER:

The user's ‘home’ directory. Set by R. (HOME will be set to the same value if not already set.)

TZDIR:

Optional. The top-level directory of the time-zone database. See Sys.timezone.

See Also

Sys.getenv and Sys.setenv to read and set environmental variables in an R session.

gctorture for environment variables controlling garbage collection.

[Package base version 4.1.2 Index]