Foreign-internal {base}	R Documentation

Internal Versions of the Foreign Function Interface

Description

Internal versions of .Call and .External.

Usage


    .Call.graphics(.NAME, ..., PACKAGE)
.External.graphics(.NAME, ..., PACKAGE)
        .External2(.NAME, ..., PACKAGE)

Arguments

.NAME
a character string giving the name of a C function or Fortran subroutine, or an object of class "NativeSymbolInfo", "RegisteredNativeSymbol" or "NativeSymbol" referring to such a name.

...
arguments to be passed to the foreign function.

PACKAGE
see the corresponding argument of .Call.

Details

The .graphics versions of .Call and .External are used when calling code which makes low-level graphics calls. They take additional steps to ensure that the device driver display lists are updated correctly.

.External2 is an enhanced version of .External which passes additional information in more arguments.

.Call.graphics is used in package grid and .External.graphics and .External2 are used in package graphics.

These are primitive functions.

Value

An R object.

See Also

.Call.

[Package base version 3.6.1 Index]