Hyperbolic {base}	R Documentation

Hyperbolic Functions

Description

These functions give the obvious hyperbolic functions. They respectively compute the hyperbolic cosine, sine, tangent, and their inverses, arc-cosine, arc-sine, arc-tangent (or “area cosine”, etc).

Usage

cosh(x)
sinh(x)
tanh(x)
acosh(x)
asinh(x)
atanh(x)

Arguments

x 	 a numeric vector

Details

These are generic functions: methods can be defined for them individually or via the Math group generic.

See Also

The trigonometric functions, cos, sin, tan, and their inverses acos, asin, atan.

The logistic distribution function plogis is a shifted version of tanh().

[Package base version 2.1.0 Index]