Quotes {base}	R Documentation

Quotes

Description

Descriptions of the various uses of quoting in R.

Details

Three types of quote are part of the syntax of R: single and double quotation marks and the backtick (or back quote, `). In addition, backslash is used for quoting the following characters inside character constants.

Character constants

Single and double quotes delimit character constants. They can be used interchangeably but double quotes are preferred (and character constants are printed using double quotes), so single quotes are normally only used to delimit character constants containing double quotes.

Backslash is used to start an escape sequence inside character constants. Unless specified in the following table, an escaped character is interpreted as the character itself. (Note that the parser will warn about most such uses, as they are most often erroneous, e.g. using \. where \\. was intended.)

Single quotes need to be escaped by backslash in single-quoted strings, and double quotes in double-quoted strings.

\n 	 newline

\r 	 carriage return

\t 	 tab

\b 	 backspace

\a 	 alert (bell)

\f 	 form feed

\v 	 vertical tab

\\ 	 backslash \

\nnn 	 character with given octal code (1, 2 or 3 digits)

\xnn 	 character with given hex code (1 or 2 hex digits)

\unnnn 	 Unicode character with given code (1–4 hex digits)

\Unnnnnnnn 	 Unicode character with given code (1–8 hex digits)

The last two are only supported on versions of R built with MBCS support: they are an error on other versions. Alternative forms are \u{nnnn} and \U{nnnnnnnn}. All except the Unicode escape sequences are also supported when reading character strings by scan and read.table if allowEscapes = TRUE.

These forms will also be used by print.default when outputting non-printable characters (including backslash).

Note that as from R 2.8.0 are not allowed by the parser in character strings, so using escapes (such as \0) for a nul will result in the string being (silently) truncated at that point.

Names and Identifiers

Identifiers consist of a sequence of letters, digits, the period (.) and the underscore. They must not start with a digit nor underscore, nor with a period followed by a digit.

The definition of a letter depends on the current locale, but only ASCII digits are considered to be digits.

Such identifiers are also known as syntactic names and may be used directly in R code. Almost always, other names can be used provided they are quoted. The preferred quote is the backtick (`), and deparse will normally use it, but under many circumstances single or double quotes can be used (as a character constant will often be converted to a name). One place where backticks may be essential is to delimit variable names in formulae: see formula.

See Also

Syntax for other aspects of the syntax.

sQuote for quoting English text.

shQuote for quoting OS commands.

The R Language Definition manual.

[Package base version 2.7.2 Index]