RdUtils {base}	R Documentation

Utilities for Processing Rd Files

Description

Utilities for converting files in R documentation (Rd) format to other formats or create indices from them, and for converting documentation in other formats to Rd format.

Usage

R CMD Rdconv [options] file
R CMD Rd2dvi [options] files
R CMD Rd2txt [options] file
R CMD Sd2Rd [options] file

Arguments

file 	 the path to a file to be processed.

files 	 a list of file names specifying the R documentation sources to use, by either giving the paths to the files, or the path to a directory with the sources of a package.

options 	 further options to control the processing, or for obtaining information about usage and version of the utility.

Details

Rdconv converts Rd format to other formats. Currently, plain text, HTML, LaTeX, S version 3 (Sd), and S version 4 (.sgml) formats are supported. It can also extract the examples for run-time testing.

Rd2dvi and Rd2txt are user-level programs for producing DVI/PDF output or pretty text output from Rd sources. Rd2dvi will make use of environment variables R_PAPERSIZE (default a4, ) and xdvi (the DVI previewer, current default yap). (Valid values for R_PAPERSIZE are a4, letter, legal and executive.)

Sd2Rd converts S (version 3 or 4) documentation formats to Rd format.

Use R CMD foo --help to obtain usage information on utility foo.

Note

Conversion to S version 3/4 formats is rough: there are some ‘.Rd’ constructs for which there is no natural analogue. They are intended as a starting point for hand-tuning.

See Also

The chapter “Processing Rd format” in the “Writing R Extensions” manual.

[Package base version 2.7.2 Index]