Sys.info {base}	R Documentation

Extract System and User Information

Description

Reports system and user information.

Usage

Sys.info()

Details

This function is not implemented on all R platforms, and returns NULL when not available. Where possible it is based on POSIX system calls. (Under Windows, it is obtained from Windows system calls.)

Sys.info() returns details of the platform R is running on, whereas R.version gives details of the platform R was built on: they may well be different.

Value

A character vector with fields

sysname 	 The operating system.

release 	 The OS release.

version 	 The OS version.

nodename 	 A name by which the machine is known on the network (if any).

machine 	 A concise description of the hardware.

login 	 The user's login name, or "unknown" if it cannot be ascertained.

user 	 The name of the real user ID, or "unknown" if it cannot be ascertained.

Note

The meaning of OS ‘release’ and ‘version’ is system-dependent and there is no guarantee that the node or login or user names will be what you might reasonably expect. (In particular on some Linux distributions the login name is unknown from sessions with re-directed inputs.)

See Also

.Platform, and R.version. sessionInfo() gives a synopsis of both your system and the R session.

Examples

Sys.info()
## An alternative (and probably better) way to get the login name on Unix
Sys.getenv("LOGNAME")

[Package base version 2.9.2 Index]