Sys.time {base}	R Documentation

Get Current Date and Time

Description

Sys.time and Sys.Date returns the system's idea of the current date with and without time.

Usage

Sys.time()
Sys.Date()

Details

Sys.time returns an absolute date-time value which can be converted in various time zones and may return different days.

Sys.Date returns the day in the current timezone.

Value

Sys.time returns an object of class "POSIXct" (see DateTimeClasses). On some systems it will have sub-second accuracy, but on others it will increment in seconds. On Windows, it increments in clock ticks (1/60 of a second) reported to millisecond accuracy.
Sys.Date returns an object of class "Date" (see Date).

See Also

date for the system time in a fixed-format character string; the elapsed time component of proc.time for possibly finer resolution in changes in time.

Sys.timezone.

Examples

Sys.time()
## print with possibly greater accuracy:
op <- options(digits.secs=6)
Sys.time()
options(op)

## locale-specific version of date()
format(Sys.time(), "%a %b %d %X %Y")

Sys.Date()

[Package base version 2.9.2 Index]