all.names {base}	R Documentation

Find All Names in an Expression

Description

Return a character vector containing all the names which occur in an expression or call.

Usage

all.names(expr, functions = TRUE, max.names = -1L, unique = FALSE)

all.vars(expr, functions = FALSE, max.names = -1L, unique = TRUE)

Arguments

expr 	 an expression or call from which the names are to be extracted.

functions 	 a logical value indicating whether function names should be included in the result.

max.names 	 the maximum number of names to be returned. -1 indicates no limit (other than vector size limits).

unique 	 a logical value which indicates whether duplicate names should be removed from the value.

Details

These functions differ only in the default values for their arguments.

Value

A character vector with the extracted names.

Examples

all.names(expression(sin(x+y)))
all.vars(expression(sin(x+y)))

[Package base version 2.9.2 Index]