args {base}	R Documentation

Argument List of a Function

Description

Displays the argument names and corresponding default values of a function.

Usage

args(name)

Arguments

name 	 an interpreted function. If name is a character string then the function with that name is found and used.

Details

This function is mainly used interactively. For programming, use formals instead.

Value

A function with identical formal argument list but an empty body if given an interpreted function; NULL in case of a variable or primitive (non-interpreted) function.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

formals, help.

Examples

args(c)            # -> NULL (c is a 'primitive' function)
args(graphics::plot.default)

[Package base version 2.1.0 Index]