args {base}	R Documentation

Argument List of a Function

Description

Displays the argument names and corresponding default values of a function or primitive.

Usage

args(name)

Arguments

name 	 a function (a closure or a primitive). If name is a character string then the function with that name is found and used.

Details

This function is mainly used interactively to print the argument list of a function. For programming, consider using formals instead.

Value

For a closure, a closure with identical formal argument list but an empty (NULL) body.
For a primitive, a closure with the documented usage and NULL body. Note that in almost all cases primitives do not make use of named arguments and match by position rather than name. (rep and seq.int are exceptions.)
NULL in case of a non-function.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

formals, help.

Examples

args(c)
args(graphics::plot.default)

[Package base version 2.9.2 Index]