base-defunct {base}	R Documentation

Defunct Functions in Package base

Description

The functions or variables listed here are no longer part of R as they are no longer needed.

Usage


# Defunct in 1.x
Version()
provide(package)
.Provided
category(...)
print.anova.glm(.)
print.anova.lm(.)
print.tabular(.)
print.plot(.)
save.plot(.)
system.test(.)
getenv(...)
read.table.url(url, method, ...)
scan.url(url, file = tempfile(), method, ...)
source.url(url, file = tempfile(), method, ...)
httpclient(url, port = 80, error.is.fatal = TRUE,
           check.MIME.type = TRUE, file = tempfile(),
           drop.ctrl.z = TRUE)
parse.dcf(text = NULL, file = "", fields = NULL,
          versionfix = FALSE)
.Alias(expr)
print.ordered(.)
.Dyn.libs
.lib.loc
machine()
Machine()
Platform()
restart()
printNoClass(x, digits = NULL, quote = TRUE, na.print = NULL,
             print.gap = NULL, right = FALSE, ...)
codes(x, ...)
codes(x, ...) <- value

# Defunct in 2.0.0
La.eigen(x, symmetric, only.values = FALSE,
         method = c("dsyevr", "dsyev"))
tetragamma(x)
pentagamma(x)
package.description(pkg, lib.loc = NULL, fields = NULL)

# Defunct in 2.2.0
loadURL(url, envir = parent.frame(), quiet = TRUE, ...)
delay(x, env = .GlobalEnv)

# Defunct in 2.3.0
format.char(x, width = NULL, flag = "-")

# Defunct in 2.4.0
La.chol(x)
La.chol2inv(x, size = ncol(x))

# Defunct in 2.4.0
symbol.C(name)
symbol.For(name)

# Defunct in 2.5.0
unix(call, intern = FALSE)

# Defunct in 2.8.0
gammaCody(x)

# Defunct in 2.9.0
manglePackageName(pkgName, pkgVersion)

# Defunct in 2.13.0
.Import(...)
.ImportFrom(name, ...)
.Export(...)
.S3method(generic, class, method)

# Defunct in 2.14.0
.First.lib(lib, pkg)

# Defunct in 2.15.0
mem.limits(nsize = NA, vsize = NA)
.saveRDS(...)
.readRDS(...)
Sys.putenv(...)

# Defunct in 3.0.0
real(length = 0)
as.real(x, ...)
is.real(x)

# Defunct in 3.1.0
.find.package(...)
.path.package(...)

Details

Some of these have stubs which report that they are defunct, but most have been removed completely (apart from being documented here).

category has been an old-S function before there were factors; should be replaced by factor throughout!

The new function dev.print() should now be used for saving plots to a file or printing them.

provide and its object .Provided have been removed. Their intended purpose was to allow one package to subsume another, but this was never used.

getenv has been replaced by Sys.getenv.

*.url are replaced by calling read.table, scan or source on a url connection.

httpclient was used by the deprecated "socket" method of download.file.

parse.dcf has been replaced by read.dcf, which is much faster, but has a slightly different interface.

.Alias provided an unreliable way to create duplicate references to the same object. There is no direct replacement. Where multiple references to a single object are required for semantic reasons consider using environments or external pointers. There are some notes on https://developer.r-project.org.

.Dyn.libs and .lib.loc were internal variables used for storing and manipulating the information about packages with DLLs, and the known R library trees. These are now dynamic variables which one can get or set using .dynLibs and .libPaths, respectively.

Machine() and Platform() were functions returning the variables .Machine and .Platform respectively.

restart() should be replaced by try(), in preparation for an exception-based implementation. If you use restart() in a way that cannot be replaced with try() then ask for help on r-devel.

printNoClass was in package methods and called directly the internal function print.default.

codes was almost always used inappropriately. To get the internal coding of a factor, use unclass, as.vector or as.integer. For ordered factors, codes was equivalent to these, but for unordered factors it assumed an alphabetical ordering of the levels in the locale in use.

La.eigen has become the default for eigen.

tetragamma and pentagamma have been replaced by psigamma.

package.description has been replaced by packageDescription in package utils.

loadURL has been superseded by load(url()).

delay has been replaced by delayedAssign. Unevaluated promises should never be visible.

write.table0 was an R-level version of write.table provided for temporary back-compatibility.

format.char was an auxiliary function for formatC whose functionality (and more) is provided by format.default.

La.chol and La.chol2inv became the default methods for chol and chol2inv.

Functions symbol.C and symbol.For were provided for historical compatibility with S. symbol.C just returns its argument unchanged, whereas symbol.For will map to lower case and append an underscore on those platforms (almost all) which do so for Fortran symbols. These are no longer of any use in R.

unix was a deprecated (since the 1990s) alternative to system, available for backwards compatibility with S.

gammaCody was the (Γ) function from the Specfun package used at C level in the Bessel code. The R-level interface was here for experimental purposes.

manglePackageName was part of the defunct support for versioned installs.

.Import, .ImportFrom, .Export and .S3method were part of an experimental alternative way to specify namespaces.

.First.lib was the load/attach hook for packages without namespaces: use .onLoad or .onAttach instead.

mem.limits reported on the old-style limits on the number of cons cells and size of the vector heap. These are no longer limited except by overall memory limits.

.readRDS and .saveRDS, were internal-only versions (and the only versions available prior to R 2.13.0) of .readRDS and .saveRDS.

Sys.putenv was a synonym for Sys.setenv, deprecated since R 2.5.0.

real, as.real and is.real were long-deprecated synonyms for the double versions.

.find.package and .path.package were internal-only versions prior to R 2.13.0, and since deprecated wrappers for find.package and path.package.

See Also

Defunct, base-deprecated

[Package base version 3.6.1 Index]