base-deprecated {base}	R Documentation

Deprecated Functions in Base package

Description

These functions are provided for compatibility with older versions of R only, and may be defunct as soon as the next release.

Usage

Sys.putenv(...)

Arguments

... 	 named arguments with values coercible to a character string.

Details

The original help page for these functions is often available at help("oldName-deprecated") (note the quotes). Functions in packages other than the base package are listed in help("pkg-deprecated").

Sys.putenv is a deprecated synonym for Sys.setenv.

See Also

Deprecated, base-defunct

[Package base version 2.9.2 Index]