base-deprecated {base}	R Documentation

Deprecated Functions in Package base

Description

These functions are provided for compatibility with older versions of R only, and may be defunct as soon as the next release.

Details

The original help page for these functions is often available at help("oldName-deprecated") (note the quotes). Functions in standard packages other than the base package are listed in help("pkg-deprecated"), where pkg is replaced by the name of the package.

See Also

Deprecated

[Package base version 3.6.1 Index]