base-deprecated {base}	R Documentation

Deprecated Functions in Base package

Description

These functions are provided for compatibility with older versions of R only, and may be defunct as soon as the next release.

Usage

Sys.putenv(...)

gammaCody(x)

Arguments

... 	 named arguments with values coercible to a character string.

x 	 numeric, >= 0.

Details

The original help page for these functions is often available at help("oldName-deprecated") (note the quotes). Functions in packages other than the base package are listed in help("pkg-deprecated").

Sys.putenv is a deprecated synonym for Sys.setenv.

gammaCody is the (Γ) function from the Specfun package and was at one time used in the Bessel code. It may be somewhat faster but less precise and/or robust than R's standard gamma. It was here for experimental purposes.

See Also

Deprecated, base-defunct

[Package base version 2.7.2 Index]