base-package {base}	R Documentation

The R Base Package

Description

Base R functions

Details

This package contains the basic functions which let R{} function as a language: arithmetic, input/output, basic programming support, etc. Its contents are available through inheritance from any environment.

For a complete list of functions, use library(help="base").

[Package base version 2.9.2 Index]