capabilities {base}	R Documentation

Report Capabilities of this Build of R

Description

Report on the optional features which have been compiled into this build of R.

Usage

capabilities(what = NULL)

Arguments

what 	 character vector or NULL, specifying required components. NULL implies that all are required.

Value

A named logical vector. Current components are

jpeg 	 Is the jpeg function operational?

png 	 Is the png function operational?

tiff 	 Is the tiff function operational?

tcltk 	 Is the tcltk package operational?

X11 	 always FALSE on Windows.

aqua 	 FALSE except on Mac OS X.

http/ftp 	 Are url and the internal method for download.file available?

sockets 	 Are make.socket and related functions available?

libxml 	 Is there support for integrating libxml with the R event loop?

fifo 	 are FIFO connections supported?

cledit 	 Is command-line editing available in the current R session? This is false in non-interactive sessions.

iconv 	 is internationalization conversion via iconv supported?

NLS 	 is there Natural Language Support (for message translations)?

profmem 	 is there support for memory profiling?

cairo 	 applies to Unix-alikes – always FALSE.

See Also

.Platform

Examples

capabilities()

if(!capabilities("http/ftp"))
   warning("internal download.file() is not available")

## See also the examples for 'connections'.

[Package base version 2.9.2 Index]