.Device {base}	R Documentation

Lists of Open/Active Graphics Devices

Description

A pairlist of the names of open graphics devices is stored in .Devices. The name of the active device (see dev.cur) is stored in .Device. Both are symbols and so appear in the base name space.

Value

.Device is a length-one character vector.
.Devices is a pairlist of length-one character vectors. The first entry is always "null device", and there are as many entries as the maximal number of graphics devices which have been simultaneously active. If a device has been removed, its entry will be "" until the device number is reused.

[Package base version 2.9.2 Index]