file.choose {base}	R Documentation

Choose a File Interactively

Description

Choose a file interactively.

Usage


file.choose(new = FALSE)

Arguments

new
Logical: choose the style of dialog box presented to the user: at present only new = FALSE is used.

Value

A character vector of length one giving the file path. This will be a marked UTF-8 string, as a Unicode file selection dialog box is used.

See Also

list.files for non-interactive selection. choose.files for selecting multiple files interactively.

[Package base version 3.6.1 Index]