file.path {base}	R Documentation

Construct Path to File

Description

Construct the path to a file from components in a platform-independent way.

Usage

file.path(..., fsep = .Platform$file.sep)

Arguments

... 	 character vectors.

fsep 	 the path separator to use.

Value

A character vector of the arguments concatenated term-by-term and separated by fsep if all arguments have positive length; otherwise, an empty character vector.

Note

The components are separated by / (not \) on Windows.

[Package base version 2.9.2 Index]