file.path {base}	R Documentation

Construct Path to File

Description

Construct the path to a file from components in a platform-independent way.

Usage


file.path(..., fsep = .Platform$file.sep)

Arguments

...
character vectors.

fsep
the path separator to use.

Details

The implementation is designed to be fast (faster than paste) as this function is used extensively in R itself.

It can also be used for environment paths such as PATH and R_LIBS with fsep = .Platform$path.sep.

Trailing path separators are invalid for Windows file paths apart from ‘/’ and ‘d:/’ (although some functions/utilities do accept them), so a trailing / or \ is removed.

Value

A character vector of the arguments concatenated term-by-term and separated by fsep if all arguments have positive length; otherwise, an empty character vector (unlike paste).

Note

The components are by default separated by / (not \) on Windows.

[Package base version 3.6.1 Index]