hexmode {base}	R Documentation

Display Numbers in Hexadecimal

Description

Convert or print integers in hexadecimal format, with as many digits as are needed to display the largest, using leading zeroes as necessary.

Usage

as.hexmode(x)

## S3 method for class 'hexmode':
as.character(x, upper.case = FALSE, ...)

## S3 method for class 'hexmode':
format(x, upper.case = FALSE, ...)

## S3 method for class 'hexmode':
print(x, ...)

Arguments

x 	 An object, for the methods inheriting from class "hexmode".

upper.case 	 a logical indicating whether to use upper-case letters or lower-case letters (default).

... 	 further arguments passed to or from other methods.

Details

Class "hexmode" consists of integer vectors with that class attribute, used merely to ensure that they are printed in hex.

See Also

octmode

[Package base version 2.9.2 Index]