is.object {base}	R Documentation

Is an Object ‘internally classed’?

Description

A function rather for internal use. It returns TRUE if the object x has the R internal OBJECT bit set, and FALSE otherwise. The OBJECT bit is set when a "class" attribute is added and removed when that attribute is removed, so this is a very efficient way to check if an object has a class attribute. (S4 objects always should.)

Usage


is.object(x)

Arguments

x
object to be tested.

Note

This is a primitive function.

See Also

class, and methods.

isS4.

Examples


is.object(1) # FALSE
is.object(as.factor(1:3)) # TRUE

[Package base version 3.6.1 Index]