is.unsorted {base}	R Documentation

Test if an Object is Not Sorted

Description

Test if an object is not sorted, without the cost of sorting it.

Usage

is.unsorted(x, na.rm = FALSE, strictly = FALSE)

Arguments

x 	 an R object with a class or a numeric, complex, character or logical vector.

na.rm 	 logical. Should missing values be removed before checking?

strictly 	 logical indicating if the check should be for strictly increasing values.

Value

A length-one logical value. All objects of length 0 or 1 are sorted: the result will be NA for objects of length 2 or more except for atomic vectors and objects with a class (where the >= or > method is used).

See Also

sort, order.

[Package base version 2.9.2 Index]