l10n_info {base}	R Documentation

Localization Information

Description

Report on localization information.

Usage

l10n_info()

Details

‘A Latin-1 locale’ includes supersets (for printable characters) such as Windows codepage 1252 but not Latin-9 (ISO 8859-15).

On Windows (where the resulting list contains codepage and system.codepage components additionally), common codepages are 1252 (Western European), 1250 (Central European), 1251 (Cyrillic), 1253 (Greek), 1254 (Turkish), 1255 (Hebrew), 1256 (Arabic), 1257 (Baltic), 1258 (Vietnamese), 874 (Thai), 932 (Japanese), 936 (Simplified Chinese), 949 (Korean) and 950 (Traditional Chinese). Codepage 28605 is Latin-9 and 65001 is UTF-8 (where supported). R does not allow the C locale, and uses 1252 as the default codepage.

Value

A list with three logical elements and further OS-specific elements:

MBCS
If a multi-byte character set in use?

UTF-8
Is this known to be a UTF-8 locale?

Latin-1
Is this known to be a Latin-1 locale?

Not on Windows:

codeset
character. The encoding name as reported by the OS, possibly "". (Added in R 4.1.0. Encoding names are OS-specific.)

Only on Windows:

codepage
integer: the Windows codepage corresponding to the locale R is using (and not necessarily that Windows is using).

system.codepage
integer: the Windows system/ANSI codepage (the codepage Windows is using). Added in R 4.1.0.

See Also

Sys.getlocale, localeconv

Examples

l10n_info()

[Package base version 4.1.2 Index]