labels {base}	R Documentation

Find Labels from Object

Description

Find a suitable set of labels from an object for use in printing or plotting, for example. A generic function.

Usage

labels(object, ...)

Arguments

object 	 Any R object: the function is generic.

... 	 further arguments passed to or from other methods.

Value

A character vector or list of such vectors. For a vector the results is the names or seq(along=x) and for a data frame or array it is the dimnames (with NULL expanded to seq(len=d[i]).

References

Chambers, J. M. and Hastie, T. J. (1992) Statistical Models in S. Wadsworth & Brooks/Cole.

[Package base version 2.9.2 Index]