lazyLoad {base}	R Documentation

Lazy Load a Database of R Objects

Description

Lazy load a database of R objects.

Usage

lazyLoad(filebase, envir = parent.frame(), filter)

Arguments

filebase 	 The file path to the database, with no extension.

envir 	 The environment into which objects are loaded.

filter 	 An optional function which when called on a a character vector of object names returns a logical vector: only objects for which this is true will be loaded.

Details

This is the workhorse function called by the package loader to load the code for a package from a database. The database consists of two binary files, ‘filebase.rdb’ (the objects) and ‘filebase.rdx’ (an index).

The objects are not themselves loaded into envir: rather promises are created that will load the object from the database on first access. (See delayedAssign.)

Author(s)

Luke Tierney

See Also

makeLazyLoading

[Package base version 2.9.2 Index]