lazyLoad {base}	R Documentation

Lazy Load a Database of R Objects

Description

Internal functions to lazy load a database of R objects.

Usage


lazyLoad(filebase, envir = parent.frame(), filter)
lazyLoadDBexec(filebase, fun, filter)

Arguments

filebase
The file path to the database, with no extension.

envir
The environment into which objects are loaded.

filter
An optional function which when called on a a character vector of object names returns a logical vector: only objects for which this is true will be loaded.

fun
Function of one argument, an environment.

Details

These are internal functions for use only by R itself.

The function lazyLoad is the workhorse function called by the package loader to load the code for a package from a database. The database consists of two binary files, ‘filebase.rdb’ (the objects) and ‘filebase.rdx’ (an index).

The objects are not themselves loaded into envir: rather promises are created that will load the object from the database on first access. (See delayedAssign.)

The function lazyLoadDBexec contains the core implementation and is also used by the mechanism for loading processed help file data.

Author(s)

Luke Tierney

[Package base version 3.6.1 Index]