license {base}	R Documentation

The R License Terms

Description

The license terms under which R is distributed.

Usage

license()
licence()

Details

R is distributed under the terms of the GNU GENERAL PUBLIC LICENSE Version 2, June 1991. A copy of this license is in file ‘R_HOME/COPYING’ and can be viewed by RShowDoc("COPYING").

A small number of files (the API header files) are distributed under the LESSER GNU GENERAL PUBLIC LICENSE version 2.1. A copy of this license is in file ‘$R_DOC_DIR/COPYING.LIB’ and can be viewed by RShowDoc("COPYING.LIB").

[Package base version 2.9.2 Index]