list.files {base}	R Documentation

List the Files in a Directory/Folder

Description

These functions produce a character vector of the names of files in the named directory.

Usage

list.files(path = ".", pattern = NULL, all.files = FALSE,
           full.names = FALSE, recursive = FALSE,
           ignore.case = FALSE)

       dir(path = ".", pattern = NULL, all.files = FALSE,
           full.names = FALSE, recursive = FALSE,
           ignore.case = FALSE)

Arguments

path 	 a character vector of full path names; the default corresponds to the working directory getwd(). Missing values will be ignored.

pattern 	 an optional regular expression. Only file names which match the regular expression will be returned.

all.files 	 a logical value. If FALSE, only the names of visible files are returned. If TRUE, all file names will be returned.

full.names 	 a logical value. If TRUE, the directory path is prepended to the file names. If FALSE, only the file names are returned.

recursive 	 logical. Should the listing recurse into directories?

ignore.case 	 logical. Should pattern-matching be case-insensitive?

Value

A character vector containing the names of the files in the specified directories, or "" if there were no files. If a path does not exist or is not a directory or is unreadable it is skipped, with a warning.
The files are sorted in alphabetical order, on the full path if full.names = TRUE. Directories are included only if recursive = FALSE.

Note

File naming conventions are platform dependent.

recursive = TRUE is not supported on all platforms and may be ignored (with a warning).

Author(s)

Ross Ihaka, Brian Ripley

See Also

file.info, file.access and files for many more file handling functions and file.choose and choose.files for interactive selection.

glob2rx to convert wildcards (as used by system file commands and shells) to regular expressions.

Sys.glob for wildcard expansion on file paths.

Examples

list.files(R.home())
## Only files starting with a-l or r
## Note that a-l is locale-dependent, but using case-insensitive
## matching makes it unambiguous in English locales
dir("../..", pattern = "^[a-lr]",full.names=TRUE, ignore.case = TRUE)

[Package base version 2.9.2 Index]