memory.profile {base}	R Documentation

Profile the Usage of Cons Cells

Description

Lists the usage of the cons cells by SEXPREC type.

Usage


memory.profile()

Details

The current types and their uses are listed in the include file ‘Rinternals.h’.

Value

A vector of counts, named by the types. See typeof for an explanation of types.

See Also

gc for the overall usage of cons cells. Rprofmem and tracemem allow memory profiling of specific code or objects, but need to be enabled at compile time.

Examples


memory.profile()

[Package base version 3.6.1 Index]