ns-internals {base}	R Documentation

Name Space Internals

Description

Internal name space support functions. Not intended to be called directly.

Usage

asNamespace(ns, base.OK = TRUE)
getNamespaceInfo(ns, which)
importIntoEnv(impenv, impnames, expenv, expnames)
isBaseNamespace(ns)
isNamespace(ns)
namespaceExport(ns, vars)
namespaceImport(self, ...)
namespaceImportFrom(self, ns, vars, generics, packages)
namespaceImportClasses(self, ns, vars)
namespaceImportMethods(self, ns, vars)
packageHasNamespace(package, package.lib)
parseNamespaceFile(package, package.lib, mustExist = TRUE)
registerS3method(genname, class, method, envir = parent.frame())
registerS3methods(info, package, env)
setNamespaceInfo(ns, which, val)
.mergeExportMethods(new, ns)
.mergeImportMethods(impenv, expenv, metaname)
.knownS3Generics

Arguments

ns 	 string or name space environment.

base.OK 	 logical.

impenv 	 environment.

expenv 	 name space environment.

vars 	 character vector.

generics 	 optional character vector.

self 	 name space environment.

package 	 string naming the package/name space to load.

packages 	 vector of package names parallel to generics.

package.lib 	 character vector specifying library.

mustExist 	 logical.

genname 	 character.

class 	 character.

envir, env 	 environment.

info 	 a 3-column character matrix.

which 	 character.

val 	 any object.

... 	 character arguments.

metaname 	 the methods table name.

Details

packageHasNamespace does not indicate if the package has a namespace, rather if it has a ‘NAMESPACE’ file, which base does not. But then you are not intended to be using it ....

Author(s)

Luke Tierney

[Package base version 2.9.2 Index]