ns-reflect.Rd {base}	R Documentation

Name Space Reflection Support

Description

Functions to support reflection on name space objects.

Usage

getExportedValue(ns, name)
getNamespace(name)
getNamespaceExports(ns)
getNamespaceImports(ns)
getNamespaceName(ns)
getNamespaceUsers(ns)
getNamespaceVersion(ns)

Arguments

ns 	 string or name space object.

name 	 string or name.

Details

getExportedValue returns the value of the exported variable name in name space ns.

getNamespace returns the environment representing the name space name. The name space is loaded if necessary.

getNamespaceExports returns a character vector of the names exported by ns.

getNamespaceImports returns a representation of the imports used by name space ns. This representation is experimental and subject to change.

getNamespaceName and getNamespaceVersion return the name and version of the name space ns.

getNamespaceUsers returns a character vector of the names of the name spaces that import name space ns.

Author(s)

Luke Tierney

[Package base version 2.9.2 Index]