ns-topenv {base}	R Documentation

Top Level Environment

Description

Finding the top level environment from an environment envir and its enclosing environments.

Usage

topenv(envir = parent.frame(),
       matchThisEnv = getOption("topLevelEnvironment"))

Arguments

envir
environment.

matchThisEnv
return this environment, if it matches before any other criterion is satisfied. The default, the option topLevelEnvironment, is set by sys.source, which treats a specific environment as the top level environment. Supplying the argument as NULL or emptyenv() means it will never match.

Details

topenv returns the first top level environment found when searching envir and its enclosing environments. If no top level environment is found, .GlobalEnv is returned. An environment is considered top level if it is the internal environment of a namespace, a package environment in the search path, or .GlobalEnv .

See Also

environment, notably parent.env() on “enclosing environments”; loadNamespace for more on namespaces.

Examples

topenv(.GlobalEnv)
topenv(new.env()) # also global env
topenv(environment(ls))# namespace:base
topenv(environment(lm))# namespace:stats

[Package base version 4.1.2 Index]