path.expand {base}	R Documentation

Expand File Paths

Description

Expand a path name, for example by replacing a leading tilde by the user's home directory (if defined on that platform).

Usage

path.expand(path)

Arguments

path
character vector containing one or more path names.

Details

On Unix - alikes:

On most builds of R a leading ~user will expand to the home directory of user (since R 4.1.0 also without readline in use).

There are possibly different concepts of ‘home directory’: that usually used is the setting of the environment variable HOME.

The ‘path names’ need not exist nor be valid path names but they do need to be representable in the session encoding.

On Windows:

The definition of the ‘home’ directory is in the ‘rw-FAQ’ Q2.14: it is taken from the R_USER environment variable when path.expand is first called in a session.

The ‘path names’ need not exist nor be valid path names.

Value

A character vector of possibly expanded path names: where the home directory is unknown or none is specified the path is returned unchanged.

See Also

basename, normalizePath, file.path.

Examples

path.expand("~/foo")

[Package base version 4.1.2 Index]