prod {base}	R Documentation

Product of Vector Elements

Description

prod returns the product of all the values present in its arguments.

Usage


prod(..., na.rm = FALSE)

Arguments

...
numeric or complex or logical vectors.

na.rm
logical. Should missing values be removed?

Details

If na.rm is FALSE an NA value in any of the arguments will cause a value of NA to be returned, otherwise NA values are ignored.

This is a generic function: methods can be defined for it directly or via the Summary group generic. For this to work properly, the arguments ... should be unnamed, and dispatch is on the first argument.

Logical true values are regarded as one, false values as zero. For historical reasons, NULL is accepted and treated as if it were numeric(0).

Value

The product, a numeric (of type "double") or complex vector of length one. NB: the product of an empty set is one, by definition.

S4 methods

This is part of the S4 Summary group generic. Methods for it must use the signature x, ..., na.rm.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

sum, cumprod, cumsum.

‘plotmath’ for the use of prod in plot annotation.

Examples


print(prod(1:7)) == print(gamma(8))

[Package base version 3.6.1 Index]