readRDS {base}	R Documentation

Internal Serialization Interface

Description

Internal functions for serialization. See serialize for public (but experimental) versions.

Usage

.saveRDS(object, file = "", ascii = FALSE, version = NULL,
         compress = TRUE, refhook = NULL)

.readRDS(file, refhook = NULL)

Arguments

object 	 R object to serialize.

file 	 a connection or the name of the file where the R object is saved to or read from.

ascii 	 a logical. If TRUE, an ASCII representation is written; otherwise (default except for text-mode connections), a more compact binary one is used.

version 	 the workspace format version to use. NULL specifies the current default format. The version used from R 0.99.0 to R 1.3.1 was version 1. The default format as from R 1.4.0 is version 2.

compress 	 a logical specifying whether saving to a named file is to use compression. Ignored when file is a connection and for workspace format version 1.

refhook 	 a hook function for handling reference objects.

Value

For .readRDS, an R object.
For .saveRDS, NULL invisibly.

[Package base version 2.9.2 Index]