real {base}	R Documentation

Real Vectors

Description

These functions are the same as their double equivalents and are provided for backwards compatibility only.

Usage

real(length = 0)
as.real(x, ...)
is.real(x)

Arguments

length 	 desired length.

x 	 object to be coerced or tested.

... 	 further arguments passed to or from other methods.

Details

as.real is a generic function, but S3 methods must be written for as.double.

as.real and is.real are primitive, so positional matching is used and any names of supplied arguments are ignored. This may not be true of methods for as.real.

[Package base version 2.9.2 Index]