search {base}	R Documentation

Give Search Path for R Objects

Description

Gives a list of attached packages (see library), and R objects, usually data.frames.

Usage

search()
searchpaths()

Value

A character vector, starting with ".GlobalEnv", and ending with "package:base" which is R's base package required always.
searchpaths gives a similar character vector, with the entries for packages being the path to the package used to load the code.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole. (search.)

Chambers, J. M. (1998) Programming with Data. A Guide to the S Language. Springer. (searchPaths.)

See Also

.packages to list just the packages on search path.

loadedNamespaces to list loaded name spaces.

attach and detach to change the search path, objects to find R objects in there.

Examples

search()
searchpaths()

[Package base version 2.9.2 Index]