shell.exec {base}	R Documentation

Open a File or URL using Windows File Associations

Description

Opens the specified file or URL using the application specified in the Windows file associations.

Usage

shell.exec(file)

Arguments

file 	 file or URL to be opened.

Details

It is likely that file needs to be a complete path, e.g. ‘"c:/R/results.html"’ or ‘"c:\R\results.html"’ as most applications interpret files relative to their working directory.

Encoded file:// URLs are not usually accepted.

Value

No value, but informative error messages will be given if the operation fails.

Author(s)

B. D. Ripley

See Also

system, shell

Examples

## Not run: 
## the space should not be encoded here
shell.exec("C:\\Program Files\\BreezeSys\\BreezeBrowser\\Breezebrowser.htm")
shell.exec("C:/Program Files/BreezeSys/BreezeBrowser/Breezebrowser.htm")
shell.exec("file://C:/Program Files/BreezeSys/BreezeBrowser/Breezebrowser.htm")
## End(Not run)

[Package base version 2.9.2 Index]