sign {base}	R Documentation

Sign Function

Description

sign returns a vector with the signs of the corresponding elements of x (the sign of a real number is 1, 0, or -1 if the number is positive, zero, or negative, respectively).

Note that sign does not operate on complex vectors.

Usage

sign(x)

Arguments

x 	 a numeric vector

Details

This is a generic function: methods can be defined for it directly or via the Math group generic.

See Also

abs

Examples

sign(pi) # == 1
sign(-2:3)# -1 -1 0 1 1 1

[Package base version 2.9.2 Index]