slotOp {base}	R Documentation

Extract Slots

Description

Extract the contents of a slot in a object with a formal (S4) class structure.

Usage

object@name

Arguments

object 	 An object from a formally defined (S4) class.

name 	 The character-string name of the slot.

Details

This operator supports the formal classes of package methods, and is enabled only when methods is loaded (as per default). See slot for further details.

It is checked that object is an S4 object (see isS4), and it is an error to attempt to use @ on any other object. (There is an exception for name .Data for internal useonly.)

If name is not a slot name, an error is thrown.

Value

The current contents of the slot.

See Also

Extract, slot

[Package base version 2.9.2 Index]