standardGeneric {base}	R Documentation

Formal Method System – Creating Generic Functions

Description

This help page is for Routines which are primitives used with the methods package. They should not be used without it.

Usage

standardGeneric(f)

Details

standardGeneric dispatches the method defined for a generic function f, using the actual arguments in the frame from which it is called.

For more, use the methods package, and see the documentation in GenericFunctions.

Author(s)

John Chambers

[Package base version 2.9.2 Index]