strtrim {base}	R Documentation

Trim Character Strings to Specified Display Widths

Description

Trim character strings to specified display widths.

Usage


strtrim(x, width)

Arguments

x
a character vector, or an object which can be coerced to a character vector by as.character.

width
Positive integer values: recycled to the length of x.

Details

‘Width’ is interpreted as the display width in a monospaced font. What happens with non-printable characters (such as backspace, tab) is implementation-dependent and may depend on the locale (e.g., they may be included in the count or they may be omitted).

Using this function rather than substr is important when there might be double-width (e.g., Chinese/Japanese/Korean) characters in the character vector.

Value

A character vector of the same length and with the same attributes as x (after possible coercion).

Elements of the result will have the encoding declared as that of the current locale (see Encoding) if the corresponding input had a declared encoding and the current locale is either Latin-1 or UTF-8.

Examples


strtrim(c("abcdef", "abcdef", "abcdef"), c(1,5,10))

[Package base version 3.6.1 Index]