structure {base}	R Documentation

Attribute Specification

Description

structure returns the given object with further attributes set.

Usage


structure(.Data, ...)

Arguments

.Data
an object which will have various attributes attached to it.

...
attributes, specified in tag = value form, which will be attached to data.

Details

Adding a class "factor" will ensure that numeric codes are given integer storage mode.

For historical reasons (these names are used when deparsing), attributes ".Dim", ".Dimnames", ".Names", ".Tsp" and ".Label" are renamed to "dim", "dimnames", "names", "tsp" and "levels".

It is possible to give the same tag more than once, in which case the last value assigned wins. As with other ways of assigning attributes, using tag = NULL removes attribute tag from .Data if it is present.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

attributes, attr.

Examples


structure(1:6, dim = 2:3)

[Package base version 3.6.1 Index]