sum {base}	R Documentation

Sum of Vector Elements

Description

sum returns the sum of all the values present in its arguments. If na.rm is FALSE an NA value in any of the arguments will cause a value of NA to be returned, otherwise NA values are ignored.

Usage

sum(..., na.rm = FALSE)

Arguments

... 	 numeric or complex vectors.

na.rm 	 logical. Should missing values be removed?

Details

This is a generic function: methods can be defined for it directly or via the Summary group generic.

Value

The sum. If all of ... are of type integer, then so is the sum, and in that case the result will be NA (with a warning) if integer overflow occurs.
NB: the sum of an empty set is zero, by definition.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

[Package base version 2.1.0 Index]