toString {base}	R Documentation

Convert an R Object to a Character String

Description

This is a helper function for format to produce a single character string describing an R object.

Usage


toString(x, ...)

## Default S3 method:
toString(x, width = NULL, ...)

Arguments

x
The object to be converted.

width
Suggestion for the maximum field width. Values of NULL or 0 indicate no maximum. The minimum value accepted is 6 and smaller values are taken as 6.

...
Optional arguments passed to or from methods.

Details

This is a generic function for which methods can be written: only the default method is described here. Most methods should honor the width argument to specify the maximum display width (as measured by nchar(type = "width")) of the result.

The default method first converts x to character and then concatenates the elements separated by ", ". If width is supplied and is not NULL, the default method returns the first width - 4 characters of the result with .... appended, if the full result would use more than width characters.

Value

A character vector of length 1 is returned.

Author(s)

Robert Gentleman

See Also

format

Examples


x <- c("a", "b", "aaaaaaaaaaa")
toString(x)
toString(x, width = 8)

[Package base version 3.6.1 Index]