unlink {base}	R Documentation

Delete Files and Directories

Description

unlink deletes the file(s) or directories specified by x.

Usage

unlink(x, recursive = FALSE)

Arguments

x 	 a character vector with the names of the file(s) or directories to be deleted. Wildcards (normally ‘*’ and ‘?’) are allowed.

recursive 	 logical. Should directories be deleted recursively?

Details

If recursive = FALSE directories are not deleted, not even empty ones.

file.remove can only remove files, but gives more detailed error information.

Value

0 for success, 1 for failure. Not deleting a non-existent file is not a failure.

Note

Prior to R version 1.2.0 the default on Unix was recursive = TRUE, and on Windows empty directories could be deleted.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

file.remove.

[Package base version 2.1.0 Index]