utf8Conversion {base}	R Documentation

Convert to or from UTF-8-encoded Character Vectors

Description

Conversion of UTF-8 encoded character vectors to and from integer vectors.

Usage

utf8ToInt(x)
intToUtf8(x, multiple = FALSE)

Arguments

x 	 object to be converted.

multiple 	 logical: should the conversion be to a single character string or multiple individual characters?

Details

These will work in any locale, including on machines that do not otherwise support multi-byte character sets.

Value

utf8ToInt converts a length-one character string encoded in UTF-8 to an integer vector of (numeric) UTF-8 code points.
intToUtf8 converts a vector of (numeric) UTF-8 code points either to a single character string or a character vector of single characters. (For a single character string 0 is silently omitted: otherwise 0 is mapped to "". Non-integral numeric values are truncated to integers.) The Encoding is declared as "UTF-8".

Examples

## Not run: 
## will only display in some locales and fonts
intToUtf8(0x03B2L) # Greek beta
## End(Not run)

[Package base version 2.9.2 Index]