which.min {base}	R Documentation

Where is the Min() or Max() ?

Description

Determines the location, i.e., index of the (first) minimum or maximum of a numeric vector.

Usage

which.min(x)
which.max(x)

Arguments

x 	 numeric (integer or double) vector, whose min or max is searched for.

Value

Missing and NaN values are discarded.
an integer of length 1 or 0 (iff x has no non-NAs), giving the index of the first minimum or maximum respectively of x.
If this extremum is unique (or empty), the results are the same as (but more efficient than) which(x == min(x)) or which(x == max(x)) respectively.

Author(s)

Martin Maechler

See Also

which, max.col, max, etc.

which.is.max in package nnet differs in breaking ties at random (and having a ‘fuzz’ in the definition of ties).

Examples

x <- c(1:4,0:5,11)
which.min(x)
which.max(x)

## it *does* work with NA's present, by discarding them:
presidents[1:30]
range(presidents, na.rm = TRUE)
which.min(presidents) # 28
which.max(presidents) #  2

[Package base version 2.9.2 Index]