withVisible {base}	R Documentation

Return both a value and its visibility

Description

This function evaluates an expression, returning it in a two element list containing its value and a flag showing whether it would automatically print.

Usage

withVisible(x)

Arguments

x 	 An expression to be evaluated.

Details

The argument is evaluated in the caller's context.

Value

value 	 The value of x after evaluation.

visible 	 logical; whether the value would auto-print.

See Also

invisible, eval

Examples

x <- 1
withVisible(x <- 1)
x
withVisible(x)

# Wrap the call in evalq() for special handling

df <- data.frame(a=1:5, b=1:5)
evalq(withVisible(a + b), envir=df)

[Package base version 2.9.2 Index]