writeLines {base}	R Documentation

Write Lines to a Connection

Description

Write text lines to a connection.

Usage

writeLines(text, con = stdout(), sep = "\n")

Arguments

text 	 A character vector

con 	 A connection object or a character string.

sep 	 character. A string to be written to the connection after each line of text.

Details

If the con is a character string, the function calls file to obtain a file connection which is opened for the duration of the function call.

If the connection is open it is written from its current position. If it is not open, it is opened for the duration of the call in "wt" mode and then closed again.

Normally writeLines is used with a text-mode connection, and the default separator is converted to the normal separator for that platform (LF on Unix/Linux, CRLF on Windows). For more control, open a binary connection and specify the precise value you want written to the file in sep. For even more control, use writeChar on a binary connection.

See Also

connections, writeChar, writeBin, readLines, cat

[Package base version 2.9.2 Index]