.Script {base}	R Documentation

Scripting Language Interface

Description

Internal interface to run a script through its interpreter.

Usage


.Script(interpreter, script, args, ...)

Arguments

interpreter
a character string naming the interpreter for the script.

script
a character string with the base file name of the script, which must be located in the ‘interpreter’ subdirectory of ‘R_HOME/share’.

args
a character string giving the arguments to pass to the script.

...
further arguments to be passed to system when invoking the interpreter on the script.

Note

This function is for R internal use only.

[Package base version 3.6.1 Index]