zpackages {base}	R Documentation

Listing of Packages

Description

.packages returns information about package availability.

Usage

.packages(all.available = FALSE, lib.loc = NULL)

Arguments

all.available 	 logical; if TRUE return a character vector of all available packages in lib.loc.

lib.loc 	 a character vector describing the location of R library trees to search through, or NULL. The default value of NULL corresponds to all libraries currently known.

Details

.packages() returns the “base names” of the currently attached packages invisibly whereas .packages(all.available = TRUE) gives (visibly) all packages available in the library location path lib.loc. For a package to be regarded as being available it must have a ‘DESCRIPTION’ file containing a valid version field.

Value

A character vector of package “base names”, invisible unless all.available = TRUE.

Author(s)

R core; Guido Masarotto for the all.available=TRUE part of .packages.

See Also

library, .libPaths.

Examples

(.packages())               # maybe just "base"
.packages(all = TRUE)       # return all available as character vector
require(splines)            # the same
(.packages())               # "splines", too
detach("package:splines")

[Package base version 2.1.0 Index]